refactor contract source into smaller files
