+++ src/contract.rs
@@ -1,47 +1,39 @@
 /// This contract implements SNIP-20 standard:
 /// https://github.com/SecretFoundation/SNIPs/blob/master/SNIP-20.md
 use cosmwasm_std::{
-    entry_point, to_binary, Addr, BankMsg, Binary, BlockInfo, CanonicalAddr, Coin, CosmosMsg,
-    Deps, DepsMut, Env, MessageInfo, Response, StdError, StdResult, Storage, Uint128, Uint64,
+    entry_point, to_binary, Binary, 
+    Deps, DepsMut, Env, MessageInfo, Response, StdError, StdResult, 
 };
 #[cfg(feature = "gas_evaporation")]
 use cosmwasm_std::Api;
-use rand_chacha::ChaChaRng;
-use rand_core::{RngCore, SeedableRng};
-use secret_toolkit::notification::{get_seed, notification_id, BloomParameters, ChannelInfoData, Descriptor, FlatDescriptor, GroupChannel, Notification, DirectChannel, StructDescriptor};
-use secret_toolkit::permit::{AllRevokedInterval, Permit, RevokedPermits, RevokedPermitsStore, TokenPermissions};
+use secret_toolkit::notification::{GroupChannel, DirectChannel,};
+use secret_toolkit::permit::{Permit, TokenPermissions};
 use secret_toolkit::utils::{pad_handle_result, pad_query_result};
 use secret_toolkit::viewing_key::{ViewingKey, ViewingKeyStore};
 use secret_toolkit_crypto::{hkdf_sha_256, sha_256, ContractPrng};
 
-use crate::batch;
+use crate::{execute, execute_admin, execute_deposit_redeem, execute_mint_burn, execute_transfer_send, query};
 
 #[cfg(feature = "gas_tracking")]
 use crate::dwb::log_dwb;
-use crate::dwb::{DelayedWriteBuffer, DWB, TX_NODES};
+use crate::dwb::{DelayedWriteBuffer, DWB};
 
-use crate::btbe::{
-    find_start_bundle, initialize_btbe, stored_balance, stored_entry, stored_tx_count
-};
+use crate::btbe::initialize_btbe;
+
 #[cfg(feature = "gas_tracking")]
 use crate::gas_tracker::{GasTracker, LoggingExt};
 #[cfg(feature = "gas_evaporation")]
 use crate::msg::Evaporator;
 use crate::msg::{
-    AllowanceGivenResult, AllowanceReceivedResult, ContractStatusLevel, ExecuteAnswer, ExecuteMsg,
-    InstantiateMsg, QueryAnswer, QueryMsg, QueryWithPermit, ResponseStatus::Success,
+    ContractStatusLevel, ExecuteMsg, InstantiateMsg, QueryAnswer, QueryMsg, QueryWithPermit,
 };
 use crate::notifications::{
-    render_group_notification, AllowanceNotification, MultiRecvdNotification, MultiSpentNotification, RecvdNotification, SpentNotification
+    AllowanceNotification, MultiRecvdNotification, MultiSpentNotification, RecvdNotification, SpentNotification
 };
-use crate::receiver::Snip20ReceiveMsg;
 use crate::state::{
-    safe_add, AllowancesStore, Config, MintersStore, ReceiverHashStore, CHANNELS, CONFIG, CONTRACT_STATUS, INTERNAL_SECRET_RELAXED, INTERNAL_SECRET_SENSITIVE, NOTIFICATIONS_ENABLED, TOTAL_SUPPLY
+    Config, MintersStore, CHANNELS, CONFIG, CONTRACT_STATUS, INTERNAL_SECRET_RELAXED, INTERNAL_SECRET_SENSITIVE, TOTAL_SUPPLY
 };
-use crate::strings::{SEND_TO_CONTRACT_ERR_MSG, TRANSFER_HISTORY_UNSUPPORTED_MSG};
-use crate::transaction_history::{
-    store_burn_action, store_deposit_action, store_mint_action, store_redeem_action, store_transfer_action, Tx
-};
+use crate::strings::TRANSFER_HISTORY_UNSUPPORTED_MSG;
 
 /// We make sure that responses from `handle` are padded to a multiple of this size.
 pub const RESPONSE_BLOCK_SIZE: usize = 256;
@@ -134,7 +126,7 @@ pub fn instantiate(
         let balance_address = deps.api.addr_canonicalize(balance.address.as_str())?;
         #[cfg(feature = "gas_tracking")]
         let mut tracker = GasTracker::new(deps.api);
-        perform_mint(
+        execute_mint_burn::perform_mint(
             deps.storage,
             &mut rng,
             &raw_admin,
@@ -210,12 +202,22 @@ pub fn execute(deps: DepsMut, env: Env, info: MessageInfo, msg: ExecuteMsg) -> S
         ContractStatusLevel::StopAll | ContractStatusLevel::StopAllButRedeems => {
             let response = match msg {
                 ExecuteMsg::SetContractStatus { level, .. } => {
-                    set_contract_status(deps, info, level)
+                    // load contract config from storage
+                    let config = CONFIG.load(deps.storage)?;
+
+                    // check that message sender is the admin
+                    if config.admin != info.sender {
+                        return Err(StdError::generic_err(
+                            "This is an admin command. Admin commands can only be run from admin address",
+                        ));
+                    }
+
+                    execute_admin::set_contract_status(deps, level)
                 }
                 ExecuteMsg::Redeem { amount, denom, .. }
                     if contract_status == ContractStatusLevel::StopAllButRedeems =>
                 {
-                    try_redeem(deps, env, info, amount, denom)
+                    execute_deposit_redeem::try_redeem(deps, env, info, amount, denom)
                 }
                 _ => Err(StdError::generic_err(
                     "This contract is stopped and this action is not allowed",
@@ -228,8 +230,8 @@ pub fn execute(deps: DepsMut, env: Env, info: MessageInfo, msg: ExecuteMsg) -> S
 
     let response = match msg.clone() {
         // Native
-        ExecuteMsg::Deposit { .. } => try_deposit(deps, env, info, &mut rng),
-        ExecuteMsg::Redeem { amount, denom, .. } => try_redeem(deps, env, info, amount, denom),
+        ExecuteMsg::Deposit { .. } => execute_deposit_redeem::try_deposit(deps, env, info, &mut rng),
+        ExecuteMsg::Redeem { amount, denom, .. } => execute_deposit_redeem::try_redeem(deps, env, info, amount, denom),
 
         // Base
         ExecuteMsg::Transfer {
@@ -237,7 +239,7 @@ pub fn execute(deps: DepsMut, env: Env, info: MessageInfo, msg: ExecuteMsg) -> S
             amount,
             memo,
             ..
-        } => try_transfer(deps, env, info, &mut rng, recipient, amount, memo),
+        } => execute_transfer_send::try_transfer(deps, env, info, &mut rng, recipient, amount, memo),
         ExecuteMsg::Send {
             recipient,
             recipient_code_hash,
@@ -245,7 +247,7 @@ pub fn execute(deps: DepsMut, env: Env, info: MessageInfo, msg: ExecuteMsg) -> S
             msg,
             memo,
             ..
-        } => try_send(
+        } => execute_transfer_send::try_send(
             deps,
             env,
             info,
@@ -257,15 +259,15 @@ pub fn execute(deps: DepsMut, env: Env, info: MessageInfo, msg: ExecuteMsg) -> S
             msg,
         ),
         ExecuteMsg::BatchTransfer { actions, .. } => {
-            try_batch_transfer(deps, env, info, &mut rng, actions)
+            execute_transfer_send::try_batch_transfer(deps, env, info, &mut rng, actions)
         }
-        ExecuteMsg::BatchSend { actions, .. } => try_batch_send(deps, env, info, &mut rng, actions),
-        ExecuteMsg::Burn { amount, memo, .. } => try_burn(deps, env, info, amount, memo),
+        ExecuteMsg::BatchSend { actions, .. } => execute_transfer_send::try_batch_send(deps, env, info, &mut rng, actions),
+        ExecuteMsg::Burn { amount, memo, .. } => execute_mint_burn::try_burn(deps, env, info, amount, memo),
         ExecuteMsg::RegisterReceive { code_hash, .. } => {
-            try_register_receive(deps, info, code_hash)
+            execute::try_register_receive(deps, info, code_hash)
         }
-        ExecuteMsg::CreateViewingKey { entropy, .. } => try_create_key(deps, env, info, entropy, &mut rng),
-        ExecuteMsg::SetViewingKey { key, .. } => try_set_key(deps, info, key),
+        ExecuteMsg::CreateViewingKey { entropy, .. } => execute::try_create_key(deps, env, info, entropy, &mut rng),
+        ExecuteMsg::SetViewingKey { key, .. } => execute::try_set_key(deps, info, key),
 
         // Allowance
         ExecuteMsg::IncreaseAllowance {
@@ -273,20 +275,20 @@ pub fn execute(deps: DepsMut, env: Env, info: MessageInfo, msg: ExecuteMsg) -> S
             amount,
             expiration,
             ..
-        } => try_increase_allowance(deps, env, info, spender, amount, expiration),
+        } => execute::try_increase_allowance(deps, env, info, spender, amount, expiration),
         ExecuteMsg::DecreaseAllowance {
             spender,
             amount,
             expiration,
             ..
-        } => try_decrease_allowance(deps, env, info, spender, amount, expiration),
+        } => execute::try_decrease_allowance(deps, env, info, spender, amount, expiration),
         ExecuteMsg::TransferFrom {
             owner,
             recipient,
             amount,
             memo,
             ..
-        } => try_transfer_from(deps, &env, info, &mut rng, owner, recipient, amount, memo),
+        } => execute_transfer_send::try_transfer_from(deps, &env, info, &mut rng, owner, recipient, amount, memo),
         ExecuteMsg::SendFrom {
             owner,
             recipient,
@@ -295,7 +297,7 @@ pub fn execute(deps: DepsMut, env: Env, info: MessageInfo, msg: ExecuteMsg) -> S
             msg,
             memo,
             ..
-        } => try_send_from(
+        } => execute_transfer_send::try_send_from(
             deps,
             env,
             &info,
@@ -308,18 +310,18 @@ pub fn execute(deps: DepsMut, env: Env, info: MessageInfo, msg: ExecuteMsg) -> S
             msg,
         ),
         ExecuteMsg::BatchTransferFrom { actions, .. } => {
-            try_batch_transfer_from(deps, &env, info, &mut rng, actions)
+            execute_transfer_send::try_batch_transfer_from(deps, &env, info, &mut rng, actions)
         }
         ExecuteMsg::BatchSendFrom { actions, .. } => {
-            try_batch_send_from(deps, env, &info, &mut rng, actions)
+            execute_transfer_send::try_batch_send_from(deps, env, &info, &mut rng, actions)
         }
         ExecuteMsg::BurnFrom {
             owner,
             amount,
             memo,
             ..
-        } => try_burn_from(deps, &env, info, owner, amount, memo),
-        ExecuteMsg::BatchBurnFrom { actions, .. } => try_batch_burn_from(deps, &env, info, actions),
+        } => execute_mint_burn::try_burn_from(deps, &env, info, owner, amount, memo),
+        ExecuteMsg::BatchBurnFrom { actions, .. } => execute_mint_burn::try_batch_burn_from(deps, &env, info, actions),
 
         // Mint
         ExecuteMsg::Mint {
@@ -327,32 +329,18 @@ pub fn execute(deps: DepsMut, env: Env, info: MessageInfo, msg: ExecuteMsg) -> S
             amount,
             memo,
             ..
-        } => try_mint(deps, env, info, &mut rng, recipient, amount, memo),
-        ExecuteMsg::BatchMint { actions, .. } => try_batch_mint(deps, env, info, &mut rng, actions),
-
-        // Other
-        ExecuteMsg::ChangeAdmin { address, .. } => change_admin(deps, info, address),
-        ExecuteMsg::SetContractStatus { level, .. } => set_contract_status(deps, info, level),
-        ExecuteMsg::AddMinters { minters, .. } => add_minters(deps, info, minters),
-        ExecuteMsg::RemoveMinters { minters, .. } => remove_minters(deps, info, minters),
-        ExecuteMsg::SetMinters { minters, .. } => set_minters(deps, info, minters),
+        } => execute_mint_burn::try_mint(deps, env, info, &mut rng, recipient, amount, memo),
+        ExecuteMsg::BatchMint { actions, .. } => execute_mint_burn::try_batch_mint(deps, env, info, &mut rng, actions),
 
         // SNIP-24
-        ExecuteMsg::RevokePermit { permit_name, .. } => revoke_permit(deps, info, permit_name),
+        ExecuteMsg::RevokePermit { permit_name, .. } => execute::revoke_permit(deps, info, permit_name),
 
         // SNIP-24.1
-        ExecuteMsg::RevokeAllPermits { interval, .. } => revoke_all_permits(deps, info, interval),
-        ExecuteMsg::DeletePermitRevocation { revocation_id, .. } => delete_permit_revocation(deps, info, revocation_id),
+        ExecuteMsg::RevokeAllPermits { interval, .. } => execute::revoke_all_permits(deps, info, interval),
+        ExecuteMsg::DeletePermitRevocation { revocation_id, .. } => execute::delete_permit_revocation(deps, info, revocation_id),
 
-        ExecuteMsg::AddSupportedDenoms { denoms, .. } => add_supported_denoms(deps, info, denoms),
-        ExecuteMsg::RemoveSupportedDenoms { denoms, .. } => {
-            remove_supported_denoms(deps, info, denoms)
-        },
-
-        // SNIP-52
-        ExecuteMsg::SetNotificationStatus { enabled, .. } => {
-            set_notification_status(deps, info, enabled)
-        },
+        // Admin functions
+        _ => admin_execute(deps, info, msg)
     };
 
     let padded_result = pad_handle_result(response, RESPONSE_BLOCK_SIZE);
@@ -363,16 +351,46 @@ pub fn execute(deps: DepsMut, env: Env, info: MessageInfo, msg: ExecuteMsg) -> S
     padded_result
 }
 
+pub fn admin_execute(deps: DepsMut, info: MessageInfo, msg: ExecuteMsg) -> StdResult<Response> {
+    // load contract config from storage
+    let mut config = CONFIG.load(deps.storage)?;
+
+    // check that message sender is the admin
+    if config.admin != info.sender {
+        return Err(StdError::generic_err(
+            "This is an admin command. Admin commands can only be run from admin address",
+        ));
+    }
+
+    match msg {
+        ExecuteMsg::ChangeAdmin { address, .. } => execute_admin::change_admin(deps, &mut config, address),
+        ExecuteMsg::SetContractStatus { level, .. } => execute_admin::set_contract_status(deps, level),
+        ExecuteMsg::AddMinters { minters, .. } => execute_admin::add_minters(deps, &config, minters),
+        ExecuteMsg::RemoveMinters { minters, .. } => execute_admin::remove_minters(deps, &config, minters),
+        ExecuteMsg::SetMinters { minters, .. } => execute_admin::set_minters(deps, &config, minters),
+        ExecuteMsg::AddSupportedDenoms { denoms, .. } => execute_admin::add_supported_denoms(deps, &mut config, denoms),
+        ExecuteMsg::RemoveSupportedDenoms { denoms, .. } => {
+            execute_admin::remove_supported_denoms(deps, &mut config, denoms)
+        },
+
+        // SNIP-52
+        ExecuteMsg::SetNotificationStatus { enabled, .. } => {
+            execute_admin::set_notification_status(deps, enabled)
+        },
+        _ => panic!("This execute type is not an admin function"),
+    }
+}
+
 #[entry_point]
 pub fn query(deps: Deps, env: Env, msg: QueryMsg) -> StdResult<Binary> {
     pad_query_result(
         match msg {
-            QueryMsg::TokenInfo {} => query_token_info(deps.storage),
-            QueryMsg::TokenConfig {} => query_token_config(deps.storage),
-            QueryMsg::ContractStatus {} => query_contract_status(deps.storage),
-            QueryMsg::ExchangeRate {} => query_exchange_rate(deps.storage),
-            QueryMsg::Minters { .. } => query_minters(deps),
-            QueryMsg::ListChannels {} => query_list_channels(deps),
+            QueryMsg::TokenInfo {} => query::query_token_info(deps.storage),
+            QueryMsg::TokenConfig {} => query::query_token_config(deps.storage),
+            QueryMsg::ContractStatus {} => query::query_contract_status(deps.storage),
+            QueryMsg::ExchangeRate {} => query::query_exchange_rate(deps.storage),
+            QueryMsg::Minters { .. } => query::query_minters(deps),
+            QueryMsg::ListChannels {} => query::query_list_channels(deps),
             QueryMsg::WithPermit { permit, query } => permit_queries(deps, env, permit, query),
 
             #[cfg(feature = "gas_tracking")]
@@ -406,7 +424,7 @@ fn permit_queries(deps: Deps, env: Env, permit: Permit, query: QueryWithPermit) 
                 )));
             }
 
-            query_balance(deps, account)
+            query::query_balance(deps, account)
         }
         QueryWithPermit::TransferHistory { .. } => {
             return Err(StdError::generic_err(TRANSFER_HISTORY_UNSUPPORTED_MSG));
@@ -419,7 +437,7 @@ fn permit_queries(deps: Deps, env: Env, permit: Permit, query: QueryWithPermit) 
                 )));
             }
 
-            query_transactions(deps, account, page.unwrap_or(0), page_size)
+            query::query_transactions(deps, account, page.unwrap_or(0), page_size)
         }
         QueryWithPermit::Allowance { owner, spender } => {
             if !permit.check_permission(&TokenPermissions::Allowance) {
@@ -436,7 +454,7 @@ fn permit_queries(deps: Deps, env: Env, permit: Permit, query: QueryWithPermit) 
                 )));
             }
 
-            query_allowance(deps, owner, spender)
+            query::query_allowance(deps, owner, spender)
         }
         QueryWithPermit::AllowancesGiven {
             owner,
@@ -459,7 +477,7 @@ fn permit_queries(deps: Deps, env: Env, permit: Permit, query: QueryWithPermit) 
                     permit.params.permissions
                 )));
             }
-            query_allowances_given(deps, account, page.unwrap_or(0), page_size)
+            query::query_allowances_given(deps, account, page.unwrap_or(0), page_size)
         }
         QueryWithPermit::AllowancesReceived {
             spender,
@@ -480,9 +498,9 @@ fn permit_queries(deps: Deps, env: Env, permit: Permit, query: QueryWithPermit) 
                     permit.params.permissions
                 )));
             }
-            query_allowances_received(deps, account, page.unwrap_or(0), page_size)
+            query::query_allowances_received(deps, account, page.unwrap_or(0), page_size)
         }
-        QueryWithPermit::ChannelInfo { channels, txhash } => query_channel_info(
+        QueryWithPermit::ChannelInfo { channels, txhash } => query::query_channel_info(
             deps,
             env,
             channels,
@@ -496,12 +514,12 @@ fn permit_queries(deps: Deps, env: Env, permit: Permit, query: QueryWithPermit) 
                     permit.params.permissions
                 )));
             } 
-            query_list_permit_revocations(deps, account.as_str()) 
+            query::query_list_permit_revocations(deps, account.as_str()) 
         },
     }
 }
 
-pub fn viewing_keys_queries(deps: Deps, env: Env,  msg: QueryMsg) -> StdResult<Binary> {
+pub fn viewing_keys_queries(deps: Deps, env: Env, msg: QueryMsg) -> StdResult<Binary> {
     let (addresses, key) = msg.get_validation_params(deps.api)?;
 
     for address in addresses {
@@ -509,7 +527,7 @@ pub fn viewing_keys_queries(deps: Deps, env: Env,  msg: QueryMsg) -> StdResult<B
         if result.is_ok() {
             return match msg {
                 // Base
-                QueryMsg::Balance { address, .. } => query_balance(deps, address),
+                QueryMsg::Balance { address, .. } => query::query_balance(deps, address),
                 QueryMsg::TransferHistory { .. } => {
                     return Err(StdError::generic_err(TRANSFER_HISTORY_UNSUPPORTED_MSG));
                 }
@@ -518,32 +536,32 @@ pub fn viewing_keys_queries(deps: Deps, env: Env,  msg: QueryMsg) -> StdResult<B
                     page,
                     page_size,
                     ..
-                } => query_transactions(deps, address, page.unwrap_or(0), page_size),
-                QueryMsg::Allowance { owner, spender, .. } => query_allowance(deps, owner, spender),
+                } => query::query_transactions(deps, address, page.unwrap_or(0), page_size),
+                QueryMsg::Allowance { owner, spender, .. } => query::query_allowance(deps, owner, spender),
                 QueryMsg::AllowancesGiven {
                     owner,
                     page,
                     page_size,
                     ..
-                } => query_allowances_given(deps, owner, page.unwrap_or(0), page_size),
+                } => query::query_allowances_given(deps, owner, page.unwrap_or(0), page_size),
                 QueryMsg::AllowancesReceived {
                     spender,
                     page,
                     page_size,
                     ..
-                } => query_allowances_received(deps, spender, page.unwrap_or(0), page_size),
+                } => query::query_allowances_received(deps, spender, page.unwrap_or(0), page_size),
                 QueryMsg::ChannelInfo {
                     channels,
                     txhash,
                     viewer,
-                } => query_channel_info(
+                } => query::query_channel_info(
                     deps,
                     env,
                     channels,
                     txhash,
                     deps.api.addr_canonicalize(viewer.address.as_str())?,
                 ),
-                QueryMsg::ListPermitRevocations{viewer, .. } => query_list_permit_revocations(deps, viewer.address.as_str()),
+                QueryMsg::ListPermitRevocations{viewer, .. } => query::query_list_permit_revocations(deps, viewer.address.as_str()),
                 _ => panic!("This query type does not require authentication"),
             };
         }
@@ -554,2680 +572,16 @@ pub fn viewing_keys_queries(deps: Deps, env: Env,  msg: QueryMsg) -> StdResult<B
     })
 }
 
-// query functions
-
-fn query_exchange_rate(storage: &dyn Storage) -> StdResult<Binary> {
-    let constants = CONFIG.load(storage)?;
-
-    if constants.deposit_is_enabled || constants.redeem_is_enabled {
-        let rate: Uint128;
-        let denom: String;
-        // if token has more decimals than SCRT, you get magnitudes of SCRT per token
-        if constants.decimals >= 6 {
-            rate = Uint128::new(10u128.pow(constants.decimals as u32 - 6));
-            denom = "SCRT".to_string();
-        // if token has less decimals, you get magnitudes token for SCRT
-        } else {
-            rate = Uint128::new(10u128.pow(6 - constants.decimals as u32));
-            denom = constants.symbol;
-        }
-        return to_binary(&QueryAnswer::ExchangeRate { rate, denom });
-    }
-    to_binary(&QueryAnswer::ExchangeRate {
-        rate: Uint128::zero(),
-        denom: String::new(),
-    })
-}
-
-fn query_token_info(storage: &dyn Storage) -> StdResult<Binary> {
-    let constants = CONFIG.load(storage)?;
-
-    let total_supply = if constants.total_supply_is_public {
-        Some(Uint128::new(TOTAL_SUPPLY.load(storage)?))
-    } else {
-        None
-    };
-
-    to_binary(&QueryAnswer::TokenInfo {
-        name: constants.name,
-        symbol: constants.symbol,
-        decimals: constants.decimals,
-        total_supply,
-    })
-}
-
-fn query_token_config(storage: &dyn Storage) -> StdResult<Binary> {
-    let constants = CONFIG.load(storage)?;
-
-    to_binary(&QueryAnswer::TokenConfig {
-        public_total_supply: constants.total_supply_is_public,
-        deposit_enabled: constants.deposit_is_enabled,
-        redeem_enabled: constants.redeem_is_enabled,
-        mint_enabled: constants.mint_is_enabled,
-        burn_enabled: constants.burn_is_enabled,
-        supported_denoms: constants.supported_denoms,
-    })
-}
-
-fn query_contract_status(storage: &dyn Storage) -> StdResult<Binary> {
-    let contract_status = CONTRACT_STATUS.load(storage)?;
-
-    to_binary(&QueryAnswer::ContractStatus {
-        status: contract_status,
-    })
-}
-
-pub fn query_transactions(
-    deps: Deps,
-    account: String,
-    page: u32,
-    page_size: u32,
-) -> StdResult<Binary> {
-    if page_size == 0 {
-        return Err(StdError::generic_err("invalid page size"));
-    }
-
-    // Notice that if query_transactions() was called by a viewing-key call, the address of
-    // 'account' has already been validated.
-    // The address of 'account' should not be validated if query_transactions() was called by a
-    // permit call, for compatibility with non-Secret addresses.
-    let account = Addr::unchecked(account);
-    let account_raw = deps.api.addr_canonicalize(account.as_str())?;
-
-    let start = page * page_size;
-    let mut end = start + page_size; // one more than end index
-
-    // first check if there are any transactions in dwb
-    let dwb = DWB.load(deps.storage)?;
-    let dwb_index = dwb.recipient_match(&account_raw);
-    let mut txs_in_dwb = vec![];
-    let txs_in_dwb_count = dwb.entries[dwb_index].list_len()?;
-    if dwb_index > 0 && txs_in_dwb_count > 0 && start < txs_in_dwb_count as u32 {
-        // skip if start is after buffer entries
-        let head_node_index = dwb.entries[dwb_index].head_node()?;
-
-        // only look if head node is not null
-        if head_node_index > 0 {
-            let head_node = TX_NODES
-                .add_suffix(&head_node_index.to_be_bytes())
-                .load(deps.storage)?;
-            txs_in_dwb = head_node.to_vec(deps.storage, deps.api)?;
-        }
-    }
-
-    //let account_slice = account_raw.as_slice();
-    let account_stored_entry = stored_entry(deps.storage, &account_raw)?;
-    let settled_tx_count = stored_tx_count(deps.storage, &account_stored_entry)?;
-    let total = txs_in_dwb_count as u32 + settled_tx_count as u32;
-    if end > total {
-        end = total;
-    }
-
-    let mut txs: Vec<Tx> = vec![];
-
-    let txs_in_dwb_count = txs_in_dwb_count as u32;
-    if start < txs_in_dwb_count && end < txs_in_dwb_count {
-        // option 1, start and end are both in dwb
-        //println!("OPTION 1");
-        txs = txs_in_dwb[start as usize..end as usize].to_vec(); // reverse chronological
-    } else if start < txs_in_dwb_count && end >= txs_in_dwb_count {
-        // option 2, start is in dwb and end is in settled txs
-        // in this case, we do not need to search for txs, just begin at last bundle and move backwards
-        //println!("OPTION 2");
-        txs = txs_in_dwb[start as usize..].to_vec(); // reverse chronological
-        let mut txs_left = (end - start).saturating_sub(txs.len() as u32);
-        if let Some(entry) = account_stored_entry {
-            let tx_bundles_idx_len = entry.history_len()?;
-            if tx_bundles_idx_len > 0 {
-                let mut bundle_idx = tx_bundles_idx_len - 1;
-                loop {
-                    let tx_bundle = entry.get_tx_bundle_at(deps.storage, bundle_idx.clone())?;
-
-                    // only look if head node is not null
-                    if tx_bundle.head_node > 0 {
-                        let head_node = TX_NODES
-                            .add_suffix(&tx_bundle.head_node.to_be_bytes())
-                            .load(deps.storage)?;
-
-                        let list_len = tx_bundle.list_len as u32;
-                        if txs_left <= list_len {
-                            txs.extend_from_slice(
-                                &head_node.to_vec(deps.storage, deps.api)?[0..txs_left as usize],
-                            );
-                            break;
-                        }
-                        txs.extend(head_node.to_vec(deps.storage, deps.api)?);
-                        txs_left = txs_left.saturating_sub(list_len);
-                    }
-                    if bundle_idx > 0 {
-                        bundle_idx -= 1;
-                    } else {
-                        break;
-                    }
-                }
-            }
-        }
-    } else if start >= txs_in_dwb_count {
-        // option 3, start is not in dwb
-        // in this case, search for where the beginning bundle is using binary search
-
-        // bundle tx offsets are chronological, but we need reverse chronological
-        // so get the settled start index as if order is reversed
-        //println!("OPTION 3");
-        let settled_start = settled_tx_count
-            .saturating_sub(start - txs_in_dwb_count)
-            .saturating_sub(1);
-
-        if let Some((bundle_idx, tx_bundle, start_at)) =
-            find_start_bundle(deps.storage, &account_raw, settled_start)?
-        {
-            let mut txs_left = end - start;
-            let list_len = tx_bundle.list_len as u32;
-            if start_at + txs_left <= list_len {
-                // only look if head node is not null
-                if tx_bundle.head_node > 0 {
-                    let head_node = TX_NODES
-                        .add_suffix(&tx_bundle.head_node.to_be_bytes())
-                        .load(deps.storage)?;
-                    // this first bundle has all the txs we need
-                    txs = head_node.to_vec(deps.storage, deps.api)?
-                        [start_at as usize..(start_at + txs_left) as usize]
-                        .to_vec();
-                }
-            } else {
-                // only look if head node is not null
-                if tx_bundle.head_node > 0 {
-                    let head_node = TX_NODES
-                        .add_suffix(&tx_bundle.head_node.to_be_bytes())
-                        .load(deps.storage)?;
-                    // get the rest of the txs in this bundle and then go back through history
-                    txs = head_node.to_vec(deps.storage, deps.api)?[start_at as usize..].to_vec();
-                    txs_left = txs_left.saturating_sub(list_len - start_at);
-                }
-
-                if bundle_idx > 0 && txs_left > 0 {
-                    // get the next earlier bundle
-                    let mut bundle_idx = bundle_idx - 1;
-                    if let Some(entry) = account_stored_entry {
-                        loop {
-                            let tx_bundle =
-                                entry.get_tx_bundle_at(deps.storage, bundle_idx.clone())?;
-                            // only look if head node is not null
-                            if tx_bundle.head_node > 0 {
-                                let head_node = TX_NODES
-                                    .add_suffix(&tx_bundle.head_node.to_be_bytes())
-                                    .load(deps.storage)?;
-                                let list_len = tx_bundle.list_len as u32;
-                                if txs_left <= list_len {
-                                    txs.extend_from_slice(
-                                        &head_node.to_vec(deps.storage, deps.api)?
-                                            [0..txs_left as usize],
-                                    );
-                                    break;
-                                }
-                                txs.extend(head_node.to_vec(deps.storage, deps.api)?);
-                                txs_left = txs_left.saturating_sub(list_len);
-                            }
-                            if bundle_idx > 0 {
-                                bundle_idx -= 1;
-                            } else {
-                                break;
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    // deterministically obfuscate ids so they are not serial to prevent metadata leak
-    let internal_secret = INTERNAL_SECRET_RELAXED.load(deps.storage)?;
-    let internal_secret_u64: u64 = u64::from_be_bytes(internal_secret[..8].try_into().unwrap());
-    let txs = txs
-        .iter()
-        .map(|tx| {
-            // PRNG(PRNG(serial_id) ^ secret)
-            let mut rng = ChaChaRng::seed_from_u64(tx.id);
-            let serial_id_rand = rng.next_u64();
-            let new_seed = serial_id_rand ^ internal_secret_u64;
-            let mut rng = ChaChaRng::seed_from_u64(new_seed);
-            let new_id = rng.next_u64() >> (64 - 53);
-            Tx {
-                id: new_id,
-                action: tx.action.clone(),
-                coins: tx.coins.clone(),
-                memo: tx.memo.clone(),
-                block_height: tx.block_height,
-                block_time: tx.block_time,
-            }
-        })
-        .collect();
-
-    let result = QueryAnswer::TransactionHistory {
-        txs,
-        total: Some(total as u64),
-    };
-    to_binary(&result)
-}
-
-pub fn query_balance(deps: Deps, account: String) -> StdResult<Binary> {
-    // Notice that if query_balance() was called by a viewing key call, the address of 'account'
-    // has already been validated.
-    // The address of 'account' should not be validated if query_balance() was called by a permit
-    // call, for compatibility with non-Secret addresses.
-    let account = Addr::unchecked(account);
-    let account = deps.api.addr_canonicalize(account.as_str())?;
-
-    let mut amount = stored_balance(deps.storage, &account)?;
-    let dwb = DWB.load(deps.storage)?;
-    let dwb_index = dwb.recipient_match(&account);
-    if dwb_index > 0 {
-        amount = amount.saturating_add(dwb.entries[dwb_index].amount()? as u128);
-    }
-    let amount = Uint128::new(amount);
-    let response = QueryAnswer::Balance { amount };
-    to_binary(&response)
-}
-
-fn query_minters(deps: Deps) -> StdResult<Binary> {
-    let minters = MintersStore::load(deps.storage)?;
-
-    let response = QueryAnswer::Minters { minters };
-    to_binary(&response)
-}
-
-// *****************
-// SNIP-52 query functions
-// *****************
-
-///
-/// ListChannels query
-///
-///   Public query to list all notification channels.
-///
-fn query_list_channels(deps: Deps) -> StdResult<Binary> {
-    let channels: Vec<String> = CHANNELS
-        .iter(deps.storage)?
-        .map(|channel| channel.unwrap())
-        .collect();
-    to_binary(&QueryAnswer::ListChannels { channels })
-}
-
-///
-/// ChannelInfo query
-///
-///   Authenticated query allows clients to obtain the seed,
-///   and Notification ID of an event for a specific tx_hash, for a specific channel.
-///
-fn query_channel_info(
-    deps: Deps,
-    env: Env,
-    channels: Vec<String>,
-    txhash: Option<String>,
-    sender_raw: CanonicalAddr,
-) -> StdResult<Binary> {
-    let secret = INTERNAL_SECRET_SENSITIVE.load(deps.storage)?;
-    let secret = secret.as_slice();
-    let seed = get_seed(&sender_raw, secret)?;
-    let mut channels_data = vec![];
-    for channel in channels {
-        let answer_id;
-        if let Some(tx_hash) = &txhash {
-            answer_id = Some(notification_id(&seed, &channel, tx_hash)?);
-        } else {
-            answer_id = None;
-        }
-        match channel.as_str() {
-            RecvdNotification::CHANNEL_ID => {
-                let channel_info_data = ChannelInfoData {
-                    mode: "txhash".to_string(),
-                    channel,
-                    answer_id,
-                    parameters: None,
-                    data: None,
-                    next_id: None,
-                    counter: None,
-                    cddl: Some(RecvdNotification::CDDL_SCHEMA.to_string()),
-                };
-                channels_data.push(channel_info_data);
-            }
-            SpentNotification::CHANNEL_ID => {
-                let channel_info_data = ChannelInfoData {
-                    mode: "txhash".to_string(),
-                    channel,
-                    answer_id,
-                    parameters: None,
-                    data: None,
-                    next_id: None,
-                    counter: None,
-                    cddl: Some(SpentNotification::CDDL_SCHEMA.to_string()),
-                };
-                channels_data.push(channel_info_data);
-            }
-            AllowanceNotification::CHANNEL_ID => {
-                let channel_info_data = ChannelInfoData {
-                    mode: "txhash".to_string(),
-                    channel,
-                    answer_id,
-                    parameters: None,
-                    data: None,
-                    next_id: None,
-                    counter: None,
-                    cddl: Some(AllowanceNotification::CDDL_SCHEMA.to_string()),
-                };
-                channels_data.push(channel_info_data);
-            }
-            MultiRecvdNotification::CHANNEL_ID => {
-                let channel_info_data = ChannelInfoData {
-                    mode: "bloom".to_string(),
-                    channel,
-                    answer_id,
-                    parameters: Some(BloomParameters {
-                        m: MultiRecvdNotification::BLOOM_M,
-                        k: MultiRecvdNotification::BLOOM_K,
-                        h: "sha256".to_string(),
-                    }),
-                    data: Some(Descriptor {
-                        r#type: format!("packet[{}]", MultiRecvdNotification::BLOOM_N),
-                        version: "1".to_string(),
-                        packet_size: MultiRecvdNotification::PACKET_SIZE as u32,
-                        data: StructDescriptor {
-                            r#type: "struct".to_string(),
-                            label: "transfer".to_string(),
-                            members: vec![
-                                FlatDescriptor {
-                                    r#type: "uint64".to_string(),
-                                    label: "flagsAndAmount".to_string(),
-                                    description: Some(
-                                        "Bit field of [0]: non-empty memo; [2]: sender is owner; [2..]: uint62 transfer amount in base denomination".to_string(),
-                                    ),
-                                },
-                                FlatDescriptor {
-                                    r#type: "bytes8".to_string(),
-                                    label: "ownerId".to_string(),
-                                    description: Some(
-                                        "The last 8 bytes of the owner's canonical address".to_string(),
-                                    ),
-                                },
-                            ],
-                        },
-                    }),
-                    counter: None,
-                    next_id: None,
-                    cddl: None,
-                };
-                channels_data.push(channel_info_data);
-            }
-            MultiSpentNotification::CHANNEL_ID => {
-                let channel_info_data = ChannelInfoData {
-                    mode: "bloom".to_string(),
-                    channel,
-                    answer_id,
-                    parameters: Some(BloomParameters {
-                        m: MultiSpentNotification::BLOOM_M,
-                        k: MultiSpentNotification::BLOOM_K,
-                        h: "sha256".to_string(),
-                    }),
-                    data: Some(Descriptor {
-                        r#type: format!("packet[{}]", MultiSpentNotification::BLOOM_N),
-                        version: "1".to_string(),
-                        packet_size: MultiSpentNotification::PACKET_SIZE as u32,
-                        data: StructDescriptor {
-                            r#type: "struct".to_string(),
-                            label: "transfer".to_string(),
-                            members: vec![
-                                FlatDescriptor {
-                                    r#type: "uint64".to_string(),
-                                    label: "flagsAndAmount".to_string(),
-                                    description: Some(
-                                        "Bit field of [0]: non-empty memo; [1]: reserved; [2..] uint62 transfer amount in base denomination".to_string(),
-                                    ),
-                                },
-                                FlatDescriptor {
-                                    r#type: "bytes8".to_string(),
-                                    label: "recipientId".to_string(),
-                                    description: Some(
-                                        "The last 8 bytes of the recipient's canonical address".to_string(),
-                                    ),
-                                },
-                                FlatDescriptor {
-                                    r#type: "uint64".to_string(),
-                                    label: "balance".to_string(),
-                                    description: Some(
-                                        "Spender's new balance after the transfer".to_string(),
-                                    ),
-                                },
-                            ],
-                        },
-                    }),
-                    counter: None,
-                    next_id: None,
-                    cddl: None,
-                };
-                channels_data.push(channel_info_data);
-            }
-            _ => {
-                return Err(StdError::generic_err(format!(
-                    "`{}` channel is undefined",
-                    channel
-                )));
-            }
-        }
-    }
-
-    to_binary(&QueryAnswer::ChannelInfo {
-        as_of_block: Uint64::from(env.block.height),
-        channels: channels_data,
-        seed,
-    })
-}
-
-// *****************
-// End SNIP-52 query functions
-// *****************
-
-// execute functions
-
-fn change_admin(deps: DepsMut, info: MessageInfo, address: String) -> StdResult<Response> {
-    let address = deps.api.addr_validate(address.as_str())?;
-
-    let mut constants = CONFIG.load(deps.storage)?;
-    check_if_admin(&constants.admin, &info.sender)?;
-
-    constants.admin = address;
-    CONFIG.save(deps.storage, &constants)?;
-
-    Ok(Response::new().set_data(to_binary(&ExecuteAnswer::ChangeAdmin { status: Success })?))
-}
-
-fn add_supported_denoms(
-    deps: DepsMut,
-    info: MessageInfo,
-    denoms: Vec<String>,
-) -> StdResult<Response> {
-    let mut config = CONFIG.load(deps.storage)?;
-
-    check_if_admin(&config.admin, &info.sender)?;
-    if !config.can_modify_denoms {
-        return Err(StdError::generic_err(
-            "Cannot modify denoms for this contract",
-        ));
-    }
-
-    for denom in denoms.iter() {
-        if !config.supported_denoms.contains(denom) {
-            config.supported_denoms.push(denom.clone());
-        }
-    }
-
-    CONFIG.save(deps.storage, &config)?;
-
-    Ok(
-        Response::new().set_data(to_binary(&ExecuteAnswer::AddSupportedDenoms {
-            status: Success,
-        })?),
-    )
-}
-
-fn remove_supported_denoms(
-    deps: DepsMut,
-    info: MessageInfo,
-    denoms: Vec<String>,
-) -> StdResult<Response> {
-    let mut config = CONFIG.load(deps.storage)?;
-
-    check_if_admin(&config.admin, &info.sender)?;
-    if !config.can_modify_denoms {
-        return Err(StdError::generic_err(
-            "Cannot modify denoms for this contract",
-        ));
-    }
-
-    for denom in denoms.iter() {
-        config.supported_denoms.retain(|x| x != denom);
-    }
-
-    CONFIG.save(deps.storage, &config)?;
-
-    Ok(
-        Response::new().set_data(to_binary(&ExecuteAnswer::RemoveSupportedDenoms {
-            status: Success,
-        })?),
-    )
-}
-
-// SNIP-52 functions
-
-fn set_notification_status(
-    deps: DepsMut,
-    info: MessageInfo,
-    enabled: bool,
-) -> StdResult<Response> {
-    let config = CONFIG.load(deps.storage)?;
-
-    check_if_admin(&config.admin, &info.sender)?;
-
-    NOTIFICATIONS_ENABLED.save(deps.storage, &enabled)?;
-    
-    Ok(
-        Response::new().set_data(to_binary(&ExecuteAnswer::SetNotificationStatus {
-            status: Success,
-        })?),
-    )
-}
-
-// end SNIP-52 functions
-
-#[allow(clippy::too_many_arguments)]
-fn try_mint_impl(
-    deps: &mut DepsMut,
-    rng: &mut ContractPrng,
-    minter: Addr,
-    recipient: Addr,
-    amount: Uint128,
-    denom: String,
-    memo: Option<String>,
-    block: &cosmwasm_std::BlockInfo,
-    #[cfg(feature = "gas_tracking")] tracker: &mut GasTracker,
-) -> StdResult<()> {
-    let raw_amount = amount.u128();
-    let raw_recipient = deps.api.addr_canonicalize(recipient.as_str())?;
-    let raw_minter = deps.api.addr_canonicalize(minter.as_str())?;
-
-    perform_mint(
-        deps.storage,
-        rng,
-        &raw_minter,
-        &raw_recipient,
-        raw_amount,
-        denom,
-        memo,
-        block,
-        #[cfg(feature = "gas_tracking")]
-        tracker,
-    )?;
-
-    Ok(())
-}
-
-#[allow(clippy::too_many_arguments)]
-fn try_mint(
-    mut deps: DepsMut,
-    env: Env,
-    info: MessageInfo,
-    rng: &mut ContractPrng,
-    recipient: String,
-    amount: Uint128,
-    memo: Option<String>,
-) -> StdResult<Response> {
-    let secret = INTERNAL_SECRET_SENSITIVE.load(deps.storage)?;
-    let secret = secret.as_slice();
-
-    let recipient = deps.api.addr_validate(recipient.as_str())?;
-
-    let constants = CONFIG.load(deps.storage)?;
-
-    if !constants.mint_is_enabled {
-        return Err(StdError::generic_err(
-            "Mint functionality is not enabled for this token.",
-        ));
-    }
-
-    let minters = MintersStore::load(deps.storage)?;
-    if !minters.contains(&info.sender) {
-        return Err(StdError::generic_err(
-            "Minting is allowed to minter accounts only",
-        ));
-    }
-
-    let mut total_supply = TOTAL_SUPPLY.load(deps.storage)?;
-    let minted_amount = safe_add(&mut total_supply, amount.u128());
-    TOTAL_SUPPLY.save(deps.storage, &total_supply)?;
-
-    #[cfg(feature = "gas_tracking")]
-    let mut tracker: GasTracker = GasTracker::new(deps.api);
-
-    let memo_len = memo.as_ref().map(|s| s.len()).unwrap_or_default();
-
-    // Note that even when minted_amount is equal to 0 we still want to perform the operations for logic consistency
-    try_mint_impl(
-        &mut deps,
-        rng,
-        info.sender,
-        recipient.clone(),
-        Uint128::new(minted_amount),
-        constants.symbol,
-        memo,
-        &env.block,
-        #[cfg(feature = "gas_tracking")]
-        &mut tracker,
-    )?;
-
-    let mut resp = Response::new()
-        .set_data(to_binary(&ExecuteAnswer::Mint { status: Success })?);
-
-    if NOTIFICATIONS_ENABLED.load(deps.storage)? {
-        let received_notification = Notification::new(
-            recipient,
-            RecvdNotification {
-                amount: minted_amount,
-                sender: None,
-                memo_len,
-                sender_is_owner: true,
-            },
-        )
-        .to_txhash_notification(deps.api, &env, secret, None)?;
-
-        resp = resp.add_attribute_plaintext(
-            received_notification.id_plaintext(),
-            received_notification.data_plaintext(),
-        );
-    }
-
-    #[cfg(feature = "gas_tracking")]
-    return Ok(resp.add_gas_tracker(tracker));
-
-    #[cfg(not(feature = "gas_tracking"))]
-    Ok(resp)
-}
-
-fn try_batch_mint(
-    mut deps: DepsMut,
-    env: Env,
-    info: MessageInfo,
-    rng: &mut ContractPrng,
-    actions: Vec<batch::MintAction>,
-) -> StdResult<Response> {
-    let secret = INTERNAL_SECRET_SENSITIVE.load(deps.storage)?;
-    let secret = secret.as_slice();
-
-    let constants = CONFIG.load(deps.storage)?;
-
-    if !constants.mint_is_enabled {
-        return Err(StdError::generic_err(
-            "Mint functionality is not enabled for this token.",
-        ));
-    }
-
-    let minters = MintersStore::load(deps.storage)?;
-    if !minters.contains(&info.sender) {
-        return Err(StdError::generic_err(
-            "Minting is allowed to minter accounts only",
-        ));
-    }
-
-    let mut total_supply = TOTAL_SUPPLY.load(deps.storage)?;
-
-    let mut notifications = vec![];
-    // Quick loop to check that the total of amounts is valid
-    for action in actions {
-        let actual_amount = safe_add(&mut total_supply, action.amount.u128());
-
-        let recipient = deps.api.addr_validate(action.recipient.as_str())?;
-
-        #[cfg(feature = "gas_tracking")]
-        let mut tracker: GasTracker = GasTracker::new(deps.api);
-
-        notifications.push(Notification::new (
-            recipient.clone(),
-            RecvdNotification {
-                amount: actual_amount,
-                sender: None,
-                memo_len: action.memo.as_ref().map(|s| s.len()).unwrap_or_default(),
-                sender_is_owner: true,
-            },
-        ));
-
-        try_mint_impl(
-            &mut deps,
-            rng,
-            info.sender.clone(),
-            recipient,
-            Uint128::new(actual_amount),
-            constants.symbol.clone(),
-            action.memo,
-            &env.block,
-            #[cfg(feature = "gas_tracking")]
-            &mut tracker,
-        )?;
-    }
-
-    TOTAL_SUPPLY.save(deps.storage, &total_supply)?;
-
-    let mut resp = Response::new()
-        .set_data(to_binary(&ExecuteAnswer::BatchMint { status: Success })?);
-    
-    if NOTIFICATIONS_ENABLED.load(deps.storage)? {
-        resp = render_group_notification(
-            deps.api,
-            MultiRecvdNotification(notifications),
-            &env.transaction.unwrap().hash,
-            env.block.random.unwrap(),
-            secret,
-            resp,
-        )?;
-    }
-
-    Ok(resp)
-}
-
-pub fn try_set_key(deps: DepsMut, info: MessageInfo, key: String) -> StdResult<Response> {
-    ViewingKey::set(deps.storage, info.sender.as_str(), key.as_str());
-    Ok(
-        Response::new().set_data(to_binary(&ExecuteAnswer::SetViewingKey {
-            status: Success,
-        })?),
-    )
-}
-
-pub fn try_create_key(
-    deps: DepsMut,
-    env: Env,
-    info: MessageInfo,
-    entropy: Option<String>,
-    rng: &mut ContractPrng,
-) -> StdResult<Response> {
-    let entropy = [entropy.unwrap_or_default().as_bytes(), &rng.rand_bytes()].concat();
-
-    let key = ViewingKey::create(
-        deps.storage,
-        &info,
-        &env,
-        info.sender.as_str(),
-        &entropy,
-    );
-
-    Ok(Response::new().set_data(to_binary(&ExecuteAnswer::CreateViewingKey { key })?))
-}
-
-fn set_contract_status(
-    deps: DepsMut,
-    info: MessageInfo,
-    status_level: ContractStatusLevel,
-) -> StdResult<Response> {
-    let constants = CONFIG.load(deps.storage)?;
-    check_if_admin(&constants.admin, &info.sender)?;
-
-    CONTRACT_STATUS.save(deps.storage, &status_level)?;
-
-    Ok(
-        Response::new().set_data(to_binary(&ExecuteAnswer::SetContractStatus {
-            status: Success,
-        })?),
-    )
-}
-
-pub fn query_allowance(deps: Deps, owner: String, spender: String) -> StdResult<Binary> {
-    // Notice that if query_allowance() was called by a viewing-key call, the addresses of 'owner'
-    // and 'spender' have already been validated.
-    // The addresses of 'owner' and 'spender' should not be validated if query_allowance() was
-    // called by a permit call, for compatibility with non-Secret addresses.
-    let owner = Addr::unchecked(owner);
-    let spender = Addr::unchecked(spender);
-
-    let allowance = AllowancesStore::load(deps.storage, &owner, &spender);
-
-    let response = QueryAnswer::Allowance {
-        owner,
-        spender,
-        allowance: Uint128::new(allowance.amount),
-        expiration: allowance.expiration,
-    };
-    to_binary(&response)
-}
-
-pub fn query_allowances_given(
-    deps: Deps,
-    owner: String,
-    page: u32,
-    page_size: u32,
-) -> StdResult<Binary> {
-    // Notice that if query_all_allowances_given() was called by a viewing-key call,
-    // the address of 'owner' has already been validated.
-    // The addresses of 'owner' should not be validated if query_all_allowances_given() was
-    // called by a permit call, for compatibility with non-Secret addresses.
-    let owner = Addr::unchecked(owner);
-
-    let all_allowances =
-        AllowancesStore::all_allowances(deps.storage, &owner, page, page_size).unwrap_or(vec![]);
-
-    let allowances_result = all_allowances
-        .into_iter()
-        .map(|(spender, allowance)| AllowanceGivenResult {
-            spender,
-            allowance: Uint128::from(allowance.amount),
-            expiration: allowance.expiration,
-        })
-        .collect();
-
-    let response = QueryAnswer::AllowancesGiven {
-        owner: owner.clone(),
-        allowances: allowances_result,
-        count: AllowancesStore::num_allowances(deps.storage, &owner),
-    };
-    to_binary(&response)
-}
-
-pub fn query_allowances_received(
-    deps: Deps,
-    spender: String,
-    page: u32,
-    page_size: u32,
-) -> StdResult<Binary> {
-    // Notice that if query_all_allowances_received() was called by a viewing-key call,
-    // the address of 'spender' has already been validated.
-    // The addresses of 'spender' should not be validated if query_all_allowances_received() was
-    // called by a permit call, for compatibility with non-Secret addresses.
-    let spender = Addr::unchecked(spender);
-
-    let all_allowed =
-        AllowancesStore::all_allowed(deps.storage, &spender, page, page_size).unwrap_or(vec![]);
-
-    let allowances = all_allowed
-        .into_iter()
-        .map(|(owner, allowance)| AllowanceReceivedResult {
-            owner,
-            allowance: Uint128::from(allowance.amount),
-            expiration: allowance.expiration,
-        })
-        .collect();
-
-    let response = QueryAnswer::AllowancesReceived {
-        spender: spender.clone(),
-        allowances,
-        count: AllowancesStore::num_allowed(deps.storage, &spender),
-    };
-    to_binary(&response)
-}
-
-fn try_deposit(
-    deps: DepsMut,
-    env: Env,
-    info: MessageInfo,
-    rng: &mut ContractPrng,
-) -> StdResult<Response> {
-    let constants = CONFIG.load(deps.storage)?;
-
-    let mut amount = Uint128::zero();
-
-    for coin in &info.funds {
-        if constants.supported_denoms.contains(&coin.denom) {
-            amount += coin.amount
-        } else {
-            return Err(StdError::generic_err(format!(
-                "Tried to deposit an unsupported coin {}",
-                coin.denom
-            )));
-        }
-    }
-
-    if amount.is_zero() {
-        return Err(StdError::generic_err("No funds were sent to be deposited"));
-    }
-
-    let mut raw_amount = amount.u128();
-
-    if !constants.deposit_is_enabled {
-        return Err(StdError::generic_err(
-            "Deposit functionality is not enabled.",
-        ));
-    }
-
-    let mut total_supply = TOTAL_SUPPLY.load(deps.storage)?;
-    raw_amount = safe_add(&mut total_supply, raw_amount);
-    TOTAL_SUPPLY.save(deps.storage, &total_supply)?;
-
-    let sender_address = deps.api.addr_canonicalize(info.sender.as_str())?;
-
-    #[cfg(feature = "gas_tracking")]
-    let mut tracker: GasTracker = GasTracker::new(deps.api);
-
-    perform_deposit(
-        deps.storage,
-        rng,
-        &sender_address,
-        raw_amount,
-        "uscrt".to_string(),
-        &env.block,
-        #[cfg(feature = "gas_tracking")]
-        &mut tracker,
-    )?;
-
-    let resp = Response::new().set_data(to_binary(&ExecuteAnswer::Deposit { status: Success })?);
-
-    #[cfg(feature = "gas_tracking")]
-    return Ok(resp.add_gas_tracker(tracker));
-
-    #[cfg(not(feature = "gas_tracking"))]
-    Ok(resp)
-}
-
-fn try_redeem(
-    deps: DepsMut,
-    env: Env,
-    info: MessageInfo,
-    amount: Uint128,
-    denom: Option<String>,
-) -> StdResult<Response> {
-    let constants = CONFIG.load(deps.storage)?;
-    if !constants.redeem_is_enabled {
-        return Err(StdError::generic_err(
-            "Redeem functionality is not enabled for this token.",
-        ));
-    }
-
-    // if denom is none and there is only 1 supported denom then we don't need to check anything
-    let withdraw_denom = if denom.is_none() && constants.supported_denoms.len() == 1 {
-        constants.supported_denoms.first().unwrap().clone()
-    // if denom is specified make sure it's on the list before trying to withdraw with it
-    } else if denom.is_some() && constants.supported_denoms.contains(denom.as_ref().unwrap()) {
-        denom.unwrap()
-    // error handling
-    } else if denom.is_none() {
-        return Err(StdError::generic_err(
-            "Tried to redeem without specifying denom, but multiple coins are supported",
-        ));
-    } else {
-        return Err(StdError::generic_err(
-            "Tried to redeem for an unsupported coin",
-        ));
-    };
-
-    let sender_address = deps.api.addr_canonicalize(info.sender.as_str())?;
-    let amount_raw = amount.u128();
-
-    let tx_id = store_redeem_action(deps.storage, amount.u128(), constants.symbol, &env.block)?;
-
-    // load delayed write buffer
-    let mut dwb = DWB.load(deps.storage)?;
-
-    #[cfg(feature = "gas_tracking")]
-    let mut tracker = GasTracker::new(deps.api);
-
-    // settle the signer's account in buffer
-    dwb.settle_sender_or_owner_account(
-        deps.storage,
-        &sender_address,
-        tx_id,
-        amount_raw,
-        "redeem",
-        false,
-        #[cfg(feature = "gas_tracking")]
-        &mut tracker,
-    )?;
-
-    DWB.save(deps.storage, &dwb)?;
-
-    let total_supply = TOTAL_SUPPLY.load(deps.storage)?;
-    if let Some(total_supply) = total_supply.checked_sub(amount_raw) {
-        TOTAL_SUPPLY.save(deps.storage, &total_supply)?;
-    } else {
-        return Err(StdError::generic_err(
-            "You are trying to redeem more tokens than what is available in the total supply",
-        ));
-    }
-
-    let token_reserve = deps
-        .querier
-        .query_balance(&env.contract.address, &withdraw_denom)?
-        .amount;
-    if amount > token_reserve {
-        return Err(StdError::generic_err(format!(
-            "You are trying to redeem for more {withdraw_denom} than the contract has in its reserve",
-        )));
-    }
-
-    let withdrawal_coins: Vec<Coin> = vec![Coin {
-        denom: withdraw_denom,
-        amount,
-    }];
-
-    let message = CosmosMsg::Bank(BankMsg::Send {
-        to_address: info.sender.clone().into_string(),
-        amount: withdrawal_coins,
-    });
-    let data = to_binary(&ExecuteAnswer::Redeem { status: Success })?;
-    let res = Response::new().add_message(message).set_data(data);
-    Ok(res)
-}
-
-#[allow(clippy::too_many_arguments)]
-fn try_transfer_impl(
-    deps: &mut DepsMut,
-    rng: &mut ContractPrng,
-    owner: &Addr,
-    recipient: &Addr,
-    amount: Uint128,
-    denom: String,
-    memo: Option<String>,
-    block: &cosmwasm_std::BlockInfo,
-    #[cfg(feature = "gas_tracking")] tracker: &mut GasTracker,
-) -> StdResult<(Notification<RecvdNotification>, Notification<SpentNotification>)> {
-    // canonicalize owner and recipient addresses
-    let raw_owner = deps.api.addr_canonicalize(owner.as_str())?;
-    let raw_recipient = deps.api.addr_canonicalize(recipient.as_str())?;
-
-    // memo length
-    let memo_len = memo.as_ref().map(|s| s.len()).unwrap_or_default();
-
-    // create the tokens received notification for recipient
-    let received_notification = Notification::new(
-        recipient.clone(),
-        RecvdNotification {
-            amount: amount.u128(),
-            sender: Some(owner.clone()),
-            memo_len,
-            sender_is_owner: true,
-        }
-    );
-
-    // perform the transfer from owner to recipient
-    let owner_balance = perform_transfer(
-        deps.storage,
-        rng,
-        &raw_owner,
-        &raw_recipient,
-        &raw_owner,
-        amount.u128(),
-        denom,
-        memo.clone(),
-        block,
-        false,
-        #[cfg(feature = "gas_tracking")]
-        tracker,
-    )?;
-
-    // create the tokens spent notification for owner
-    let spent_notification = Notification::new (
-        owner.clone(),
-        SpentNotification {
-            amount: amount.u128(),
-            actions: 1,
-            recipient: Some(recipient.clone()),
-            balance: owner_balance,
-            memo_len,
-        }
-    );
-
-    Ok((received_notification, spent_notification))
-}
-
-#[allow(clippy::too_many_arguments)]
-fn try_transfer(
-    mut deps: DepsMut,
-    env: Env,
-    info: MessageInfo,
-    rng: &mut ContractPrng,
-    recipient: String,
-    amount: Uint128,
-    memo: Option<String>,
-) -> StdResult<Response> {
-    let secret = INTERNAL_SECRET_SENSITIVE.load(deps.storage)?;
-    let secret = secret.as_slice();
-
-    let recipient: Addr = deps.api.addr_validate(recipient.as_str())?;
-
-    let symbol = CONFIG.load(deps.storage)?.symbol;
-
-    // make sure the sender is not accidentally sending tokens to the contract address
-    if recipient == env.contract.address {
-        return Err(StdError::generic_err(SEND_TO_CONTRACT_ERR_MSG));
-    }
-
-    #[cfg(feature = "gas_tracking")]
-    let mut tracker: GasTracker = GasTracker::new(deps.api);
-
-    // perform the transfer
-    let (
-        received_notification,
-        spent_notification
-    ) = try_transfer_impl(
-        &mut deps,
-        rng,
-        &info.sender,
-        &recipient,
-        amount,
-        symbol,
-        memo,
-        &env.block,
-        #[cfg(feature = "gas_tracking")]
-        &mut tracker,
-    )?;
-
-    #[cfg(feature = "gas_tracking")]
-    let mut group1 = tracker.group("try_transfer.rest");
-
-    let mut resp = Response::new()
-        .set_data(to_binary(&ExecuteAnswer::Transfer { status: Success })?);
-
-    if NOTIFICATIONS_ENABLED.load(deps.storage)? {
-        // render the tokens received notification
-        let received_notification = received_notification.to_txhash_notification(
-            deps.api,
-            &env,
-            secret,
-            None,
-        )?;
-
-        // render the tokens spent notification
-        let spent_notification = spent_notification.to_txhash_notification(
-            deps.api, 
-            &env, 
-            secret,
-            None,
-        )?;
-        
-        resp = resp.add_attribute_plaintext(
-            received_notification.id_plaintext(),
-            received_notification.data_plaintext(),
-        )
-        .add_attribute_plaintext(
-            spent_notification.id_plaintext(),
-            spent_notification.data_plaintext(),
-        );
-    }
-
-    #[cfg(feature = "gas_tracking")]
-    group1.log("rest");
-
-    #[cfg(feature = "gas_tracking")]
-    return Ok(resp.add_gas_tracker(tracker));
-
-    #[cfg(not(feature = "gas_tracking"))]
-    Ok(resp)
-}
-
-fn try_batch_transfer(
-    mut deps: DepsMut,
-    env: Env,
-    info: MessageInfo,
-    rng: &mut ContractPrng,
-    actions: Vec<batch::TransferAction>,
-) -> StdResult<Response> {
-    let num_actions = actions.len();
-    if num_actions == 0 {
-        return Ok(Response::new()
-            .set_data(to_binary(&ExecuteAnswer::BatchTransfer { status: Success })?)
-        );
-    }
-
-    let secret = INTERNAL_SECRET_SENSITIVE.load(deps.storage)?;
-    let secret = secret.as_slice();
-
-    let symbol = CONFIG.load(deps.storage)?.symbol;
-
-    let mut total_memo_len = 0;
-
-    #[cfg(feature = "gas_tracking")]
-    let mut tracker: GasTracker = GasTracker::new(deps.api);
-
-    let mut notifications = vec![];
-    for action in actions {
-        let recipient = deps.api.addr_validate(action.recipient.as_str())?;
-      
-        // make sure the sender is not accidentally sending tokens to the contract address
-        if recipient == env.contract.address {
-            return Err(StdError::generic_err(SEND_TO_CONTRACT_ERR_MSG));
-        }
-
-        total_memo_len += action.memo.as_ref().map(|s| s.len()).unwrap_or_default();
-
-        let (
-            received_notification,
-            spent_notification
-        ) = try_transfer_impl(
-            &mut deps,
-            rng,
-            &info.sender,
-            &recipient,
-            action.amount,
-            symbol.clone(),
-            action.memo,
-            &env.block,
-            #[cfg(feature = "gas_tracking")]
-            &mut tracker,
-        )?;
-
-        notifications.push((received_notification, spent_notification));
-    }
-
-    let (
-        received_notifications,
-        spent_notifications
-    ): (
-        Vec<Notification<RecvdNotification>>,
-        Vec<Notification<SpentNotification>>,
-    ) = notifications.into_iter().unzip();
-
-    let mut resp = Response::new()
-        .set_data(to_binary(&ExecuteAnswer::BatchTransfer { status: Success })?);
-
-    if NOTIFICATIONS_ENABLED.load(deps.storage)? {
-        resp = render_group_notification(
-            deps.api,
-            MultiRecvdNotification(received_notifications),
-            &env.transaction.clone().unwrap().hash,
-            env.block.random.clone().unwrap(),
-            secret,
-            resp,
-        )?;
-
-        let total_amount_spent = spent_notifications
-            .iter()
-            .fold(0u128, |acc, notification| acc.saturating_add(notification.data.amount));
-
-        let spent_notification = Notification::new (
-            info.sender,
-            SpentNotification {
-                amount: total_amount_spent,
-                actions: num_actions as u32,
-                recipient: spent_notifications[0].data.recipient.clone(),
-                balance: spent_notifications.last().unwrap().data.balance,
-                memo_len: total_memo_len,
-            }
-        )
-        .to_txhash_notification(deps.api, &env, secret, None)?;
-
-        resp = resp.add_attribute_plaintext(
-            spent_notification.id_plaintext(),
-            spent_notification.data_plaintext(),
-        );
-    }
-
-    #[cfg(feature = "gas_tracking")]
-    return Ok(resp.add_gas_tracker(tracker));
-
-    #[cfg(not(feature = "gas_tracking"))]
-    Ok(resp)
-}
-
-#[allow(clippy::too_many_arguments)]
-fn try_add_receiver_api_callback(
-    storage: &dyn Storage,
-    messages: &mut Vec<CosmosMsg>,
-    recipient: Addr,
-    recipient_code_hash: Option<String>,
-    msg: Option<Binary>,
-    sender: Addr,
-    from: Addr,
-    amount: Uint128,
-    memo: Option<String>,
-) -> StdResult<()> {
-    if let Some(receiver_hash) = recipient_code_hash {
-        let receiver_msg = Snip20ReceiveMsg::new(sender, from, amount, memo, msg);
-        let callback_msg = receiver_msg.into_cosmos_msg(receiver_hash, recipient)?;
-
-        messages.push(callback_msg);
-        return Ok(());
-    }
-
-    let receiver_hash = ReceiverHashStore::may_load(storage, &recipient)?;
-    if let Some(receiver_hash) = receiver_hash {
-        let receiver_msg = Snip20ReceiveMsg::new(sender, from, amount, memo, msg);
-        let callback_msg = receiver_msg.into_cosmos_msg(receiver_hash, recipient)?;
-
-        messages.push(callback_msg);
-    }
-    Ok(())
-}
-
-#[allow(clippy::too_many_arguments)]
-fn try_send_impl(
-    deps: &mut DepsMut,
-    rng: &mut ContractPrng,
-    messages: &mut Vec<CosmosMsg>,
-    sender: Addr,
-    recipient: Addr,
-    recipient_code_hash: Option<String>,
-    amount: Uint128,
-    denom: String,
-    memo: Option<String>,
-    msg: Option<Binary>,
-    block: &cosmwasm_std::BlockInfo,
-    #[cfg(feature = "gas_tracking")] tracker: &mut GasTracker,
-) -> StdResult<(Notification<RecvdNotification>, Notification<SpentNotification>)> {
-    let (
-        received_notification,
-        spent_notification
-    ) = try_transfer_impl(
-        deps,
-        rng,
-        &sender,
-        &recipient,
-        amount,
-        denom,
-        memo.clone(),
-        block,
-        #[cfg(feature = "gas_tracking")]
-        tracker,
-    )?;
-
-    try_add_receiver_api_callback(
-        deps.storage,
-        messages,
-        recipient,
-        recipient_code_hash,
-        msg,
-        sender.clone(),
-        sender,
-        amount,
-        memo,
-    )?;
-
-    Ok((received_notification, spent_notification))
-}
-
-#[allow(clippy::too_many_arguments)]
-fn try_send(
-    mut deps: DepsMut,
-    env: Env,
-    info: MessageInfo,
-    rng: &mut ContractPrng,
-    recipient: String,
-    recipient_code_hash: Option<String>,
-    amount: Uint128,
-    memo: Option<String>,
-    msg: Option<Binary>,
-) -> StdResult<Response> {
-    let secret = INTERNAL_SECRET_SENSITIVE.load(deps.storage)?;
-    let secret = secret.as_slice();
-
-    let recipient = deps.api.addr_validate(recipient.as_str())?;
-
-    let mut messages = vec![];
-    let symbol = CONFIG.load(deps.storage)?.symbol;
-
-    // make sure the sender is not accidentally sending tokens to the contract address
-    if recipient == env.contract.address {
-        return Err(StdError::generic_err(SEND_TO_CONTRACT_ERR_MSG));
-    }
-
-    #[cfg(feature = "gas_tracking")]
-    let mut tracker: GasTracker = GasTracker::new(deps.api);
-
-    let (
-        received_notification,
-        spent_notification
-    ) = try_send_impl(
-        &mut deps,
-        rng,
-        &mut messages,
-        info.sender,
-        recipient,
-        recipient_code_hash,
-        amount,
-        symbol,
-        memo,
-        msg,
-        &env.block,
-        #[cfg(feature = "gas_tracking")]
-        &mut tracker,
-    )?;
-
-    let mut resp = Response::new()
-        .add_messages(messages)
-        .set_data(to_binary(&ExecuteAnswer::Send { status: Success })?);
-
-    if NOTIFICATIONS_ENABLED.load(deps.storage)? {
-        let received_notification = received_notification.to_txhash_notification(deps.api, &env, secret, None)?;
-        let spent_notification = spent_notification.to_txhash_notification(deps.api, &env, secret, None)?;
-            
-        resp = resp.add_attribute_plaintext(
-            received_notification.id_plaintext(),
-            received_notification.data_plaintext(),
-        )
-        .add_attribute_plaintext(
-            spent_notification.id_plaintext(),
-            spent_notification.data_plaintext(),
-        );
-    }
-
-    #[cfg(feature = "gas_tracking")]
-    return Ok(resp.add_gas_tracker(tracker));
-
-    #[cfg(not(feature = "gas_tracking"))]
-    Ok(resp)
-}
-
-fn try_batch_send(
-    mut deps: DepsMut,
-    env: Env,
-    info: MessageInfo,
-    rng: &mut ContractPrng,
-    actions: Vec<batch::SendAction>,
-) -> StdResult<Response> {
-    let num_actions = actions.len();
-    if num_actions == 0 {
-        return Ok(Response::new()
-            .set_data(to_binary(&ExecuteAnswer::BatchSend { status: Success })?)
-        );
-    }
-
-    let secret = INTERNAL_SECRET_SENSITIVE.load(deps.storage)?;
-    let secret = secret.as_slice();
-
-    let mut messages = vec![];
-
-    let mut notifications = vec![];
-    let num_actions: usize = actions.len();
-
-    let symbol = CONFIG.load(deps.storage)?.symbol;
-
-    let mut total_memo_len = 0;
-
-    #[cfg(feature = "gas_tracking")]
-    let mut tracker: GasTracker = GasTracker::new(deps.api);
-
-    for action in actions {
-        let recipient = deps.api.addr_validate(action.recipient.as_str())?;
-
-        // make sure the sender is not accidentally sending tokens to the contract address
-        if recipient == env.contract.address {
-            return Err(StdError::generic_err(SEND_TO_CONTRACT_ERR_MSG));
-        }
-
-        total_memo_len += action.memo.as_ref().map(|s| s.len()).unwrap_or_default();
-
-        let (
-            received_notification,
-            spent_notification
-        ) = try_send_impl(
-            &mut deps,
-            rng,
-            &mut messages,
-            info.sender.clone(),
-            recipient,
-            action.recipient_code_hash,
-            action.amount,
-            symbol.clone(),
-            action.memo,
-            action.msg,
-            &env.block,
-            #[cfg(feature = "gas_tracking")]
-            &mut tracker,
-        )?;
-
-        notifications.push((received_notification, spent_notification));
-    }
-
-    let mut resp = Response::new()
-        .add_messages(messages)
-        .set_data(to_binary(&ExecuteAnswer::BatchSend { status: Success })?);
-
-    if NOTIFICATIONS_ENABLED.load(deps.storage)? {
-        let (received_notifications, spent_notifications): (
-            Vec<Notification<RecvdNotification>>,
-            Vec<Notification<SpentNotification>>,
-        ) = notifications.into_iter().unzip();
-
-        resp = render_group_notification(
-            deps.api,
-            MultiRecvdNotification(received_notifications),
-            &env.transaction.clone().unwrap().hash,
-            env.block.random.clone().unwrap(),
-            secret,
-            resp,
-        )?;
-
-        let total_amount_spent = spent_notifications
-            .iter()
-            .fold(0u128, |acc, notification| acc + notification.data.amount);
-
-        let spent_notification = Notification::new (
-            info.sender,
-            SpentNotification {
-                amount: total_amount_spent,
-                actions: num_actions as u32,
-                recipient: spent_notifications[0].data.recipient.clone(),
-                balance: spent_notifications.last().unwrap().data.balance,
-                memo_len: total_memo_len,
-            }
-        )
-        .to_txhash_notification(deps.api, &env, secret, None)?;
-
-        resp = resp.add_attribute_plaintext(
-            spent_notification.id_plaintext(),
-            spent_notification.data_plaintext(),
-        );
-    }
-
-    Ok(resp)
-}
-
-fn try_register_receive(
-    deps: DepsMut,
-    info: MessageInfo,
-    code_hash: String,
-) -> StdResult<Response> {
-    ReceiverHashStore::save(deps.storage, &info.sender, code_hash)?;
-
-    let data = to_binary(&ExecuteAnswer::RegisterReceive { status: Success })?;
-    Ok(Response::new()
-        .add_attribute("register_status", "success")
-        .set_data(data))
-}
-
-fn insufficient_allowance(allowance: u128, required: u128) -> StdError {
-    StdError::generic_err(format!(
-        "insufficient allowance: allowance={allowance}, required={required}",
-    ))
-}
-
-fn use_allowance(
-    storage: &mut dyn Storage,
-    env: &Env,
-    owner: &Addr,
-    spender: &Addr,
-    amount: u128,
-) -> StdResult<()> {
-    let mut allowance = AllowancesStore::load(storage, owner, spender);
-
-    if allowance.is_expired_at(&env.block) || allowance.amount == 0 {
-        return Err(insufficient_allowance(0, amount));
-    }
-    if let Some(new_allowance) = allowance.amount.checked_sub(amount) {
-        allowance.amount = new_allowance;
-    } else {
-        return Err(insufficient_allowance(allowance.amount, amount));
-    }
-
-    AllowancesStore::save(storage, owner, spender, &allowance)?;
-
-    Ok(())
-}
-
-#[allow(clippy::too_many_arguments)]
-fn try_transfer_from_impl(
-    deps: &mut DepsMut,
-    rng: &mut ContractPrng,
-    env: &Env,
-    spender: &Addr,
-    owner: &Addr,
-    recipient: &Addr,
-    amount: Uint128,
-    denom: String,
-    memo: Option<String>,
-) -> StdResult<(Notification<RecvdNotification>, Notification<SpentNotification>)> {
-    let raw_amount = amount.u128();
-    let raw_spender = deps.api.addr_canonicalize(spender.as_str())?;
-    let raw_owner = deps.api.addr_canonicalize(owner.as_str())?;
-    let raw_recipient = deps.api.addr_canonicalize(recipient.as_str())?;
-
-    use_allowance(deps.storage, env, owner, spender, raw_amount)?;
-
-    // make sure the sender is not accidentally sending tokens to the contract address
-    if *recipient == env.contract.address {
-        return Err(StdError::generic_err(SEND_TO_CONTRACT_ERR_MSG));
-    }
-
-    #[cfg(feature = "gas_tracking")]
-    let mut tracker: GasTracker = GasTracker::new(deps.api);
-
-    let memo_len = memo.as_ref().map(|s| s.len()).unwrap_or_default();
-
-    // create tokens received notification for recipient
-    let received_notification = Notification::new(
-        recipient.clone(),
-        RecvdNotification {
-            amount: amount.u128(),
-            sender: Some(owner.clone()),
-            memo_len,
-            sender_is_owner: spender == owner,
-        }
-    );
-    
-    // perform the transfer from owner to recipient
-    let owner_balance = perform_transfer(
-        deps.storage,
-        rng,
-        &raw_owner,
-        &raw_recipient,
-        &raw_spender,
-        raw_amount,
-        denom,
-        memo,
-        &env.block,
-        true,
-        #[cfg(feature = "gas_tracking")]
-        &mut tracker,
-    )?;
-
-    // create tokens spent notification for owner
-    let spent_notification = Notification::new (
-        owner.clone(),
-        SpentNotification {
-            amount: amount.u128(),
-            actions: 1,
-            recipient: Some(recipient.clone()),
-            balance: owner_balance,
-            memo_len,
-        }
-    );
-
-    Ok((received_notification, spent_notification))
-}
-
-#[allow(clippy::too_many_arguments)]
-fn try_transfer_from(
-    mut deps: DepsMut,
-    env: &Env,
-    info: MessageInfo,
-    rng: &mut ContractPrng,
-    owner: String,
-    recipient: String,
-    amount: Uint128,
-    memo: Option<String>,
-) -> StdResult<Response> {
-    let secret = INTERNAL_SECRET_SENSITIVE.load(deps.storage)?;
-    let secret = secret.as_slice();
-
-    let owner = deps.api.addr_validate(owner.as_str())?;
-    let recipient = deps.api.addr_validate(recipient.as_str())?;
-    let symbol = CONFIG.load(deps.storage)?.symbol;
-    let (
-        received_notification,
-        spent_notification
-    ) = try_transfer_from_impl(
-        &mut deps,
-        rng,
-        env,
-        &info.sender,
-        &owner,
-        &recipient,
-        amount,
-        symbol,
-        memo,
-    )?;
-
-    let mut resp = Response::new()
-        .set_data(to_binary(&ExecuteAnswer::TransferFrom { status: Success })?);
-
-    if NOTIFICATIONS_ENABLED.load(deps.storage)? {
-        let received_notification = received_notification.to_txhash_notification(
-            deps.api,
-            &env,
-            secret,
-            None,
-        )?;
-
-        let spent_notification = spent_notification.to_txhash_notification(
-            deps.api, 
-            &env, 
-            secret, 
-            None
-        )?;
-
-        resp = resp.add_attribute_plaintext(
-            received_notification.id_plaintext(),
-            received_notification.data_plaintext(),
-        )
-        .add_attribute_plaintext(
-            spent_notification.id_plaintext(),
-            spent_notification.data_plaintext(),
-        );
-    }
-
-    Ok(resp)
-}
-
-fn try_batch_transfer_from(
-    mut deps: DepsMut,
-    env: &Env,
-    info: MessageInfo,
-    rng: &mut ContractPrng,
-    actions: Vec<batch::TransferFromAction>,
-) -> StdResult<Response> {
-    let secret = INTERNAL_SECRET_SENSITIVE.load(deps.storage)?;
-    let secret = secret.as_slice();
-
-    let mut notifications = vec![];
-
-    let symbol = CONFIG.load(deps.storage)?.symbol;
-    for action in actions {
-        let owner = deps.api.addr_validate(action.owner.as_str())?;
-        let recipient = deps.api.addr_validate(action.recipient.as_str())?;
-
-        let (
-            received_notification,
-            spent_notification
-        ) = try_transfer_from_impl(
-            &mut deps,
-            rng,
-            env,
-            &info.sender,
-            &owner,
-            &recipient,
-            action.amount,
-            symbol.clone(),
-            action.memo,
-        )?;
-
-        notifications.push((received_notification, spent_notification));
-    }
-
-    let mut resp = Response::new()
-        .set_data(to_binary(&ExecuteAnswer::BatchTransferFrom {status: Success})?);
-
-    if NOTIFICATIONS_ENABLED.load(deps.storage)? {
-        let (received_notifications, spent_notifications): (
-            Vec<Notification<RecvdNotification>>,
-            Vec<Notification<SpentNotification>>,
-        ) = notifications.into_iter().unzip();
-
-        let tx_hash = env.transaction.clone().unwrap().hash;
-
-        resp = render_group_notification(
-            deps.api,
-            MultiRecvdNotification(received_notifications),
-            &tx_hash,
-            env.block.random.clone().unwrap(),
-            secret,
-            resp,
-        )?;
-    
-        resp = render_group_notification(
-            deps.api,
-            MultiSpentNotification(spent_notifications),
-            &tx_hash,
-            env.block.random.clone().unwrap(),
-            secret,
-            resp,
-        )?;
-    }
-
-    Ok(resp)
-}
-
-#[allow(clippy::too_many_arguments)]
-fn try_send_from_impl(
-    deps: &mut DepsMut,
-    env: Env,
-    info: &MessageInfo,
-    rng: &mut ContractPrng,
-    messages: &mut Vec<CosmosMsg>,
-    owner: Addr,
-    recipient: Addr,
-    recipient_code_hash: Option<String>,
-    amount: Uint128,
-    memo: Option<String>,
-    msg: Option<Binary>,
-) -> StdResult<(Notification<RecvdNotification>, Notification<SpentNotification>)> {
-    let spender = info.sender.clone();
-    let symbol = CONFIG.load(deps.storage)?.symbol;
-    let (
-        received_notification,
-        spent_notification
-    ) = try_transfer_from_impl(
-        deps,
-        rng,
-        &env,
-        &spender,
-        &owner,
-        &recipient,
-        amount,
-        symbol,
-        memo.clone(),
-    )?;
-
-    try_add_receiver_api_callback(
-        deps.storage,
-        messages,
-        recipient,
-        recipient_code_hash,
-        msg,
-        info.sender.clone(),
-        owner,
-        amount,
-        memo,
-    )?;
-
-    Ok((received_notification, spent_notification))
-}
-
-#[allow(clippy::too_many_arguments)]
-fn try_send_from(
-    mut deps: DepsMut,
-    env: Env,
-    info: &MessageInfo,
-    rng: &mut ContractPrng,
-    owner: String,
-    recipient: String,
-    recipient_code_hash: Option<String>,
-    amount: Uint128,
-    memo: Option<String>,
-    msg: Option<Binary>,
-) -> StdResult<Response> {
-    let secret = INTERNAL_SECRET_SENSITIVE.load(deps.storage)?;
-    let secret = secret.as_slice();
-
-    let owner = deps.api.addr_validate(owner.as_str())?;
-    let recipient = deps.api.addr_validate(recipient.as_str())?;
-    let mut messages = vec![];
-    let (
-        received_notification,
-        spent_notification
-    ) = try_send_from_impl(
-        &mut deps,
-        env.clone(),
-        info,
-        rng,
-        &mut messages,
-        owner,
-        recipient,
-        recipient_code_hash,
-        amount,
-        memo,
-        msg,
-    )?;
-
-    let mut resp = Response::new()
-        .add_messages(messages)
-        .set_data(to_binary(&ExecuteAnswer::SendFrom { status: Success })?);
-
-    if NOTIFICATIONS_ENABLED.load(deps.storage)? {
-        let received_notification = received_notification.to_txhash_notification(deps.api, &env, secret, None,)?;
-        let spent_notification = spent_notification.to_txhash_notification(deps.api, &env, secret, None)?;
-    
-        resp = resp.add_attribute_plaintext(
-            received_notification.id_plaintext(),
-            received_notification.data_plaintext(),
-        )
-        .add_attribute_plaintext(
-            spent_notification.id_plaintext(),
-            spent_notification.data_plaintext(),
-        )
-    }
-
-    Ok(resp)
-}
-
-fn try_batch_send_from(
-    mut deps: DepsMut,
-    env: Env,
-    info: &MessageInfo,
-    rng: &mut ContractPrng,
-    actions: Vec<batch::SendFromAction>,
-) -> StdResult<Response> {
-    let secret = INTERNAL_SECRET_SENSITIVE.load(deps.storage)?;
-    let secret = secret.as_slice();
-
-    let mut messages = vec![];
-    let mut notifications = vec![];
-
-    for action in actions {
-        let owner = deps.api.addr_validate(action.owner.as_str())?;
-        let recipient = deps.api.addr_validate(action.recipient.as_str())?;
-        let (
-            received_notification,
-            spent_notification
-        ) = try_send_from_impl(
-            &mut deps,
-            env.clone(),
-            info,
-            rng,
-            &mut messages,
-            owner,
-            recipient,
-            action.recipient_code_hash,
-            action.amount,
-            action.memo,
-            action.msg,
-        )?;
-        notifications.push((received_notification, spent_notification));
-    }
-
-    let mut resp = Response::new()
-        .add_messages(messages)
-        .set_data(to_binary(&ExecuteAnswer::BatchSendFrom {
-            status: Success,
-        })?);
-
-    if NOTIFICATIONS_ENABLED.load(deps.storage)? {
-        let (received_notifications, spent_notifications): (
-            Vec<Notification<RecvdNotification>>,
-            Vec<Notification<SpentNotification>>,
-        ) = notifications.into_iter().unzip();
-
-        let tx_hash = env.transaction.clone().unwrap().hash;
-
-        resp = render_group_notification(
-            deps.api,
-            MultiRecvdNotification(received_notifications),
-            &tx_hash,
-            env.block.random.clone().unwrap(),
-            secret,
-            resp,
-        )?;
-
-        resp = render_group_notification(
-            deps.api,
-            MultiSpentNotification(spent_notifications),
-            &tx_hash,
-            env.block.random.clone().unwrap(),
-            secret,
-            resp,
-        )?;
-    }
-
-    Ok(resp)
-}
-
-#[allow(clippy::too_many_arguments)]
-fn try_burn_from(
-    deps: DepsMut,
-    env: &Env,
-    info: MessageInfo,
-    owner: String,
-    amount: Uint128,
-    memo: Option<String>,
-) -> StdResult<Response> {
-    let secret = INTERNAL_SECRET_SENSITIVE.load(deps.storage)?;
-    let secret = secret.as_slice();
-    
-    let owner = deps.api.addr_validate(owner.as_str())?;
-    let raw_owner = deps.api.addr_canonicalize(owner.as_str())?;
-    let constants = CONFIG.load(deps.storage)?;
-    if !constants.burn_is_enabled {
-        return Err(StdError::generic_err(
-            "Burn functionality is not enabled for this token.",
-        ));
-    }
-
-    let raw_amount = amount.u128();
-    use_allowance(deps.storage, env, &owner, &info.sender, raw_amount)?;
-    let raw_burner = deps.api.addr_canonicalize(info.sender.as_str())?;
-
-    let memo_len = memo.as_ref().map(|s| s.len()).unwrap_or_default();
-
-    // store the event
-    let tx_id = store_burn_action(
-        deps.storage,
-        raw_owner.clone(),
-        raw_burner.clone(),
-        raw_amount,
-        constants.symbol,
-        memo,
-        &env.block,
-    )?;
-
-    // load delayed write buffer
-    let mut dwb = DWB.load(deps.storage)?;
-
-    #[cfg(feature = "gas_tracking")]
-    let mut tracker = GasTracker::new(deps.api);
-
-    // settle the owner's account in buffer
-    let owner_balance = dwb.settle_sender_or_owner_account(
-        deps.storage,
-        &raw_owner,
-        tx_id,
-        raw_amount,
-        "burn",
-        raw_burner == raw_owner,
-        #[cfg(feature = "gas_tracking")]
-        &mut tracker,
-    )?;
-
-    // sender and owner are different
-    if raw_burner != raw_owner {
-        // also settle sender's account
-        dwb.settle_sender_or_owner_account(
-            deps.storage,
-            &raw_burner,
-            tx_id,
-            0,
-            "burn",
-            false,
-            #[cfg(feature = "gas_tracking")]
-            &mut tracker,
-        )?;
-    }
-
-    DWB.save(deps.storage, &dwb)?;
-
-    // remove from supply
-    let mut total_supply = TOTAL_SUPPLY.load(deps.storage)?;
-
-    if let Some(new_total_supply) = total_supply.checked_sub(raw_amount) {
-        total_supply = new_total_supply;
-    } else {
-        return Err(StdError::generic_err(
-            "You're trying to burn more than is available in the total supply",
-        ));
-    }
-
-    TOTAL_SUPPLY.save(deps.storage, &total_supply)?;
-
-    let mut resp =  Response::new()
-        .set_data(to_binary(&ExecuteAnswer::BurnFrom { status: Success })?);
-
-    if NOTIFICATIONS_ENABLED.load(deps.storage)? {
-        let spent_notification = Notification::new (
-            owner,
-            SpentNotification {
-                amount: raw_amount,
-                actions: 1,
-                recipient: None,
-                balance: owner_balance,
-                memo_len,
-            }
-        )
-        .to_txhash_notification(deps.api, &env, secret, None)?;
-
-        resp = resp.add_attribute_plaintext(
-            spent_notification.id_plaintext(),
-            spent_notification.data_plaintext(),
-        );
-    }
-
-    Ok(resp)
-}
-
-fn try_batch_burn_from(
-    deps: DepsMut,
-    env: &Env,
-    info: MessageInfo,
-    actions: Vec<batch::BurnFromAction>,
-) -> StdResult<Response> {
-    let secret = INTERNAL_SECRET_SENSITIVE.load(deps.storage)?;
-    let secret = secret.as_slice();
-
-    let constants = CONFIG.load(deps.storage)?;
-    if !constants.burn_is_enabled {
-        return Err(StdError::generic_err(
-            "Burn functionality is not enabled for this token.",
-        ));
-    }
-
-    let raw_spender = deps.api.addr_canonicalize(info.sender.as_str())?;
-    let mut total_supply = TOTAL_SUPPLY.load(deps.storage)?;
-    let mut spent_notifications = vec![];
-
-    for action in actions {
-        let owner = deps.api.addr_validate(action.owner.as_str())?;
-        let raw_owner = deps.api.addr_canonicalize(owner.as_str())?;
-        let amount = action.amount.u128();
-        use_allowance(deps.storage, env, &owner, &info.sender, amount)?;
-
-        let tx_id = store_burn_action(
-            deps.storage,
-            raw_owner.clone(),
-            raw_spender.clone(),
-            amount,
-            constants.symbol.clone(),
-            action.memo.clone(),
-            &env.block,
-        )?;
-
-        // load delayed write buffer
-        let mut dwb = DWB.load(deps.storage)?;
-
-        #[cfg(feature = "gas_tracking")]
-        let mut tracker = GasTracker::new(deps.api);
-
-        // settle the owner's account in buffer
-        let owner_balance = dwb.settle_sender_or_owner_account(
-            deps.storage,
-            &raw_owner,
-            tx_id,
-            amount,
-            "burn",
-            raw_spender == raw_owner,
-            #[cfg(feature = "gas_tracking")]
-            &mut tracker,
-        )?;
-
-        // sender and owner are different
-        if raw_spender != raw_owner {
-            // also settle the sender's account
-            dwb.settle_sender_or_owner_account(
-                deps.storage,
-                &raw_spender,
-                tx_id,
-                0,
-                "burn",
-                false,
-                #[cfg(feature = "gas_tracking")]
-                &mut tracker,
-            )?;
-        }
-
-        DWB.save(deps.storage, &dwb)?;
-
-        // remove from supply
-        if let Some(new_total_supply) = total_supply.checked_sub(amount) {
-            total_supply = new_total_supply;
-        } else {
-            return Err(StdError::generic_err(format!(
-                "You're trying to burn more than is available in the total supply: {action:?}",
-            )));
-        }
-
-        spent_notifications.push(Notification::new (
-            info.sender.clone(),
-            SpentNotification {
-                amount,
-                actions: 1,
-                recipient: None,
-                balance: owner_balance,
-                memo_len: action.memo.as_ref().map(|s| s.len()).unwrap_or_default()
-            }
-        ));
-    }
-
-    TOTAL_SUPPLY.save(deps.storage, &total_supply)?;
-
-    let mut resp = Response::new()
-        .set_data(to_binary(&ExecuteAnswer::BatchBurnFrom {status: Success,})?);
-
-    if NOTIFICATIONS_ENABLED.load(deps.storage)? {
-        resp = render_group_notification(
-            deps.api,
-            MultiSpentNotification(spent_notifications),
-            &env.transaction.clone().unwrap().hash,
-            env.block.random.clone().unwrap(),
-            secret,
-            resp,
-        )?;
-    }
-
-    Ok(resp)
-}
-
-fn try_increase_allowance(
-    deps: DepsMut,
-    env: Env,
-    info: MessageInfo,
-    spender: String,
-    amount: Uint128,
-    expiration: Option<u64>,
-) -> StdResult<Response> {
-    let secret = INTERNAL_SECRET_SENSITIVE.load(deps.storage)?;
-    let secret = secret.as_slice();
-
-    let spender = deps.api.addr_validate(spender.as_str())?;
-    let mut allowance = AllowancesStore::load(deps.storage, &info.sender, &spender);
-
-    // If the previous allowance has expired, reset the allowance.
-    // Without this users can take advantage of an expired allowance given to
-    // them long ago.
-    if allowance.is_expired_at(&env.block) {
-        allowance.amount = amount.u128();
-        allowance.expiration = None;
-    } else {
-        allowance.amount = allowance.amount.saturating_add(amount.u128());
-    }
-
-    if expiration.is_some() {
-        allowance.expiration = expiration;
-    }
-    let new_amount = allowance.amount;
-    AllowancesStore::save(deps.storage, &info.sender, &spender, &allowance)?;
-
-    let mut resp = Response::new()
-        .set_data(to_binary(&ExecuteAnswer::IncreaseAllowance {
-            owner: info.sender.clone(),
-            spender: spender.clone(),
-            allowance: Uint128::from(new_amount),
-        })?);
-    
-    if NOTIFICATIONS_ENABLED.load(deps.storage)? {
-        let notification = Notification::new (
-            spender,
-            AllowanceNotification {
-                amount: new_amount,
-                allower: info.sender,
-                expiration,
-            }
-        )
-        .to_txhash_notification(deps.api, &env, secret, None)?;
-
-        resp = resp.add_attribute_plaintext(
-            notification.id_plaintext(),
-            notification.data_plaintext()
-        );
-    }
-
-    Ok(resp)  
-}
-
-fn try_decrease_allowance(
-    deps: DepsMut,
-    env: Env,
-    info: MessageInfo,
-    spender: String,
-    amount: Uint128,
-    expiration: Option<u64>,
-) -> StdResult<Response> {
-    let secret = INTERNAL_SECRET_SENSITIVE.load(deps.storage)?;
-    let secret = secret.as_slice();
-
-    let spender = deps.api.addr_validate(spender.as_str())?;
-    let mut allowance = AllowancesStore::load(deps.storage, &info.sender, &spender);
-
-    // If the previous allowance has expired, reset the allowance.
-    // Without this users can take advantage of an expired allowance given to
-    // them long ago.
-    if allowance.is_expired_at(&env.block) {
-        allowance.amount = 0;
-        allowance.expiration = None;
-    } else {
-        allowance.amount = allowance.amount.saturating_sub(amount.u128());
-    }
-
-    if expiration.is_some() {
-        allowance.expiration = expiration;
-    }
-    let new_amount = allowance.amount;
-    AllowancesStore::save(deps.storage, &info.sender, &spender, &allowance)?;
-
-    let mut resp = Response::new()
-        .set_data(to_binary(&ExecuteAnswer::DecreaseAllowance {
-            owner: info.sender.clone(),
-            spender: spender.clone(),
-            allowance: Uint128::from(new_amount),
-        })?);
-
-    if NOTIFICATIONS_ENABLED.load(deps.storage)? {
-        let notification = Notification::new (
-            spender,
-            AllowanceNotification {
-                amount: new_amount,
-                allower: info.sender,
-                expiration,
-            }
-        )
-        .to_txhash_notification(deps.api, &env, secret, None)?;
-
-        resp = resp.add_attribute_plaintext(
-            notification.id_plaintext(),
-            notification.data_plaintext()
-        );
-    }
-
-    Ok(resp)
-}
-
-fn add_minters(
-    deps: DepsMut,
-    info: MessageInfo,
-    minters_to_add: Vec<String>,
-) -> StdResult<Response> {
-    let constants = CONFIG.load(deps.storage)?;
-    if !constants.mint_is_enabled {
-        return Err(StdError::generic_err(
-            "Mint functionality is not enabled for this token.",
-        ));
-    }
-
-    check_if_admin(&constants.admin, &info.sender)?;
-
-    let minters_to_add: Vec<Addr> = minters_to_add
-        .iter()
-        .map(|minter| deps.api.addr_validate(minter.as_str()).unwrap())
-        .collect();
-    MintersStore::add_minters(deps.storage, minters_to_add)?;
-
-    Ok(Response::new().set_data(to_binary(&ExecuteAnswer::AddMinters { status: Success })?))
-}
-
-fn remove_minters(
-    deps: DepsMut,
-    info: MessageInfo,
-    minters_to_remove: Vec<String>,
-) -> StdResult<Response> {
-    let constants = CONFIG.load(deps.storage)?;
-    if !constants.mint_is_enabled {
-        return Err(StdError::generic_err(
-            "Mint functionality is not enabled for this token.",
-        ));
-    }
-
-    check_if_admin(&constants.admin, &info.sender)?;
-
-    let minters_to_remove: StdResult<Vec<Addr>> = minters_to_remove
-        .iter()
-        .map(|minter| deps.api.addr_validate(minter.as_str()))
-        .collect();
-    MintersStore::remove_minters(deps.storage, minters_to_remove?)?;
-
-    Ok(
-        Response::new().set_data(to_binary(&ExecuteAnswer::RemoveMinters {
-            status: Success,
-        })?),
-    )
-}
-
-fn set_minters(
-    deps: DepsMut,
-    info: MessageInfo,
-    minters_to_set: Vec<String>,
-) -> StdResult<Response> {
-    let constants = CONFIG.load(deps.storage)?;
-    if !constants.mint_is_enabled {
-        return Err(StdError::generic_err(
-            "Mint functionality is not enabled for this token.",
-        ));
-    }
-
-    check_if_admin(&constants.admin, &info.sender)?;
-
-    let minters_to_set: Vec<Addr> = minters_to_set
-        .iter()
-        .map(|minter| deps.api.addr_validate(minter.as_str()).unwrap())
-        .collect();
-    MintersStore::save(deps.storage, minters_to_set)?;
-
-    Ok(Response::new().set_data(to_binary(&ExecuteAnswer::SetMinters { status: Success })?))
-}
-
-/// Burn tokens
-///
-/// Remove `amount` tokens from the system irreversibly, from signer account
-///
-/// @param amount the amount of money to burn
-fn try_burn(
-    deps: DepsMut,
-    env: Env,
-    info: MessageInfo,
-    amount: Uint128,
-    memo: Option<String>,
-) -> StdResult<Response> {
-    let secret = INTERNAL_SECRET_SENSITIVE.load(deps.storage)?;
-    let secret = secret.as_slice();
-
-    let constants = CONFIG.load(deps.storage)?;
-    if !constants.burn_is_enabled {
-        return Err(StdError::generic_err(
-            "Burn functionality is not enabled for this token.",
-        ));
-    }
-
-    let raw_amount = amount.u128();
-    let raw_burn_address = deps.api.addr_canonicalize(info.sender.as_str())?;
-
-    let memo_len = memo.as_ref().map(|s| s.len()).unwrap_or_default();
-
-    let tx_id = store_burn_action(
-        deps.storage,
-        raw_burn_address.clone(),
-        raw_burn_address.clone(),
-        raw_amount,
-        constants.symbol,
-        memo,
-        &env.block,
-    )?;
-
-    // load delayed write buffer
-    let mut dwb = DWB.load(deps.storage)?;
-
-    #[cfg(feature = "gas_tracking")]
-    let mut tracker = GasTracker::new(deps.api);
-
-    // settle the signer's account in buffer
-    let owner_balance = dwb.settle_sender_or_owner_account(
-        deps.storage,
-        &raw_burn_address,
-        tx_id,
-        raw_amount,
-        "burn",
-        false,
-        #[cfg(feature = "gas_tracking")]
-        &mut tracker,
-    )?;
-
-    DWB.save(deps.storage, &dwb)?;
-
-    let mut total_supply = TOTAL_SUPPLY.load(deps.storage)?;
-    if let Some(new_total_supply) = total_supply.checked_sub(raw_amount) {
-        total_supply = new_total_supply;
-    } else {
-        return Err(StdError::generic_err(
-            "You're trying to burn more than is available in the total supply",
-        ));
-    }
-    TOTAL_SUPPLY.save(deps.storage, &total_supply)?;
-
-    let mut resp = Response::new()
-        .set_data(to_binary(&ExecuteAnswer::Burn { status: Success })?);
-
-    if NOTIFICATIONS_ENABLED.load(deps.storage)? {
-        let spent_notification = Notification::new (
-            info.sender,
-            SpentNotification {
-                amount: raw_amount,
-                actions: 1,
-                recipient: None,
-                balance: owner_balance,
-                memo_len,
-            }
-        )
-        .to_txhash_notification(deps.api, &env, secret, None)?;
-
-        resp = resp.add_attribute_plaintext(
-            spent_notification.id_plaintext(),
-            spent_notification.data_plaintext(),
-        );
-    }
-
-    Ok(resp)
-}
-
-fn perform_transfer(
-    store: &mut dyn Storage,
-    rng: &mut ContractPrng,
-    from: &CanonicalAddr,
-    to: &CanonicalAddr,
-    sender: &CanonicalAddr,
-    amount: u128,
-    denom: String,
-    memo: Option<String>,
-    block: &BlockInfo,
-    is_from_action: bool,
-    #[cfg(feature = "gas_tracking")] tracker: &mut GasTracker,
-) -> StdResult<u128> {
-    #[cfg(feature = "gas_tracking")]
-    let mut group1 = tracker.group("perform_transfer.1");
-
-    // first store the tx information in the global append list of txs and get the new tx id
-    let tx_id = store_transfer_action(store, from, sender, to, amount, denom, memo, block)?;
-
-    #[cfg(feature = "gas_tracking")]
-    group1.log("@store_transfer_action");
-
-    // load delayed write buffer
-    let mut dwb = DWB.load(store)?;
-
-    #[cfg(feature = "gas_tracking")]
-    group1.log("DWB.load");
-
-    let transfer_str = "transfer";
-
-    // settle the owner's account
-    let owner_balance = dwb.settle_sender_or_owner_account(
-        store,
-        from,
-        tx_id,
-        amount,
-        transfer_str,
-        is_from_action && sender == from,
-        #[cfg(feature = "gas_tracking")]
-        tracker,
-    )?;
-
-    // sender and owner are different
-    if sender != from {
-        // settle the sender's account too
-        dwb.settle_sender_or_owner_account(
-            store,
-            sender,
-            tx_id,
-            0,
-            transfer_str,
-            false,
-            #[cfg(feature = "gas_tracking")]
-            tracker,
-        )?;
-    }
-
-    // add the tx info for the recipient to the buffer
-    dwb.add_recipient(
-        store,
-        rng,
-        to,
-        tx_id,
-        amount,
-        #[cfg(feature = "gas_tracking")]
-        tracker,
-    )?;
-
-    #[cfg(feature = "gas_tracking")]
-    let mut group2 = tracker.group("perform_transfer.2");
-
-    DWB.save(store, &dwb)?;
-
-    #[cfg(feature = "gas_tracking")]
-    group2.log("DWB.save");
-
-    Ok(owner_balance)
-}
-
-fn perform_mint(
-    store: &mut dyn Storage,
-    rng: &mut ContractPrng,
-    minter: &CanonicalAddr,
-    to: &CanonicalAddr,
-    amount: u128,
-    denom: String,
-    memo: Option<String>,
-    block: &BlockInfo,
-    #[cfg(feature = "gas_tracking")] tracker: &mut GasTracker,
-) -> StdResult<()> {
-    // first store the tx information in the global append list of txs and get the new tx id
-    let tx_id = store_mint_action(store, minter, to, amount, denom, memo, block)?;
-
-    // load delayed write buffer
-    let mut dwb = DWB.load(store)?;
-
-    // sender and owner are different
-    if minter != to {
-        // settle the sender's account too
-        dwb.settle_sender_or_owner_account(
-            store,
-            minter,
-            tx_id,
-            0,
-            "mint",
-            false,
-            #[cfg(feature = "gas_tracking")]
-            tracker,
-        )?;
-    }
-
-    // add the tx info for the recipient to the buffer
-    dwb.add_recipient(
-        store,
-        rng,
-        to,
-        tx_id,
-        amount,
-        #[cfg(feature = "gas_tracking")]
-        tracker,
-    )?;
-
-    DWB.save(store, &dwb)?;
-
-    Ok(())
-}
-
-fn perform_deposit(
-    store: &mut dyn Storage,
-    rng: &mut ContractPrng,
-    to: &CanonicalAddr,
-    amount: u128,
-    denom: String,
-    block: &BlockInfo,
-    #[cfg(feature = "gas_tracking")] tracker: &mut GasTracker,
-) -> StdResult<()> {
-    // first store the tx information in the global append list of txs and get the new tx id
-    let tx_id = store_deposit_action(store, amount, denom, block)?;
-
-    // load delayed write buffer
-    let mut dwb = DWB.load(store)?;
-
-    // add the tx info for the recipient to the buffer
-    dwb.add_recipient(
-        store,
-        rng,
-        to,
-        tx_id,
-        amount,
-        #[cfg(feature = "gas_tracking")]
-        tracker,
-    )?;
-
-    DWB.save(store, &dwb)?;
-
-    Ok(())
-}
-
-fn revoke_permit(deps: DepsMut, info: MessageInfo, permit_name: String) -> StdResult<Response> {
-    RevokedPermits::revoke_permit(
-        deps.storage,
-        info.sender.as_str(),
-        &permit_name,
-    );
-
-    Ok(Response::new().set_data(to_binary(&ExecuteAnswer::RevokePermit { status: Success })?))
-}
-
-// SNIP 24.1
-
-fn revoke_all_permits(deps: DepsMut, info: MessageInfo, interval: AllRevokedInterval) -> StdResult<Response> {
-    let revocation_id = RevokedPermits::revoke_all_permits(
-        deps.storage,
-        info.sender.as_str(),
-        &interval,
-    )?;
-
-    Ok(Response::new().set_data(to_binary(&ExecuteAnswer::RevokeAllPermits { 
-        status: Success,
-        revocation_id: Some(revocation_id.to_string()),
-    })?))
-}
-
-fn delete_permit_revocation(deps: DepsMut, info: MessageInfo, revocation_id: String) -> StdResult<Response> {
-    RevokedPermits::delete_revocation(
-        deps.storage,
-        info.sender.as_str(),
-        revocation_id.as_str(),
-    )?;
-
-    Ok(Response::new().set_data(to_binary(&ExecuteAnswer::DeletePermitRevocation { 
-        status: Success,
-    })?))
-}
-
-fn query_list_permit_revocations(deps: Deps, account: &str) -> StdResult<Binary> {
-    let revocations = RevokedPermits::list_revocations(
-        deps.storage, 
-        account
-    )?;
-
-    to_binary(&QueryAnswer::ListPermitRevocations { revocations })
-}
-
-// end SNIP 24.1
-
-fn check_if_admin(config_admin: &Addr, account: &Addr) -> StdResult<()> {
-    if config_admin != account {
-        return Err(StdError::generic_err(
-            "This is an admin command. Admin commands can only be run from admin address",
-        ));
-    }
-
-    Ok(())
-}
+// pub fn migrate(
+//     _deps: DepsMut,
+//     _env: Env,
+//     _msg: MigrateMsg,
+// ) -> StdResult<MigrateResponse> {
+//     Ok(MigrateResponse::default())
+//     Ok(MigrateResponse::default())
+// }
+
+// helper functions
 
 fn is_valid_name(name: &str) -> bool {
     let len = name.len();
@@ -3241,29 +595,22 @@ fn is_valid_symbol(symbol: &str) -> bool {
     len_is_valid && symbol.bytes().all(|byte| byte.is_ascii_alphabetic())
 }
 
-// pub fn migrate(
-//     _deps: DepsMut,
-//     _env: Env,
-//     _msg: MigrateMsg,
-// ) -> StdResult<MigrateResponse> {
-//     Ok(MigrateResponse::default())
-//     Ok(MigrateResponse::default())
-// }
-
 #[cfg(test)]
 mod tests {
     use std::any::Any;
 
     use cosmwasm_std::{
-        from_binary, testing::*, Api, BlockInfo, ContractInfo, MessageInfo, OwnedDeps,
-        QueryResponse, ReplyOn, SubMsg, Timestamp, TransactionInfo, WasmMsg,
+        from_binary, testing::*, Addr, Api, BlockInfo, Coin, ContractInfo, CosmosMsg, MessageInfo, OwnedDeps, QueryResponse, ReplyOn, SubMsg, Timestamp, TransactionInfo, Uint128, WasmMsg
     };
     use secret_toolkit::permit::{PermitParams, PermitSignature, PubKey};
 
-    use crate::dwb::TX_NODES_COUNT;
-    use crate::msg::{InitConfig, InitialBalance, ResponseStatus};
-    use crate::state::TX_COUNT;
-    use crate::transaction_history::TxAction;
+    use crate::batch;
+    use crate::btbe::stored_balance;
+    use crate::dwb::{TX_NODES, TX_NODES_COUNT};
+    use crate::msg::{ExecuteAnswer, InitConfig, InitialBalance, ResponseStatus, ResponseStatus::Success};
+    use crate::receiver::Snip20ReceiveMsg;
+    use crate::state::{AllowancesStore, ReceiverHashStore, TX_COUNT};
+    use crate::transaction_history::{Tx, TxAction};
 
     use super::*;
