+++ src/lib.rs
@@ -6,6 +6,12 @@ mod btbe;
 mod constants;
 pub mod contract;
 mod dwb;
+pub mod execute;
+pub mod execute_admin;
+pub mod execute_deposit_redeem;
+pub mod execute_mint_burn;
+pub mod execute_transfer_send;
+pub mod query;
 mod gas_tracker;
 pub mod msg;
 pub mod receiver;
